backtest_metrics

Compute metrics for given backtest results

Usage

backtest_metrics(..., metrics = c("mae", "rmse", "wis"), horizons = 0)

Arguments

...

Results of calls to backtest.

metrics

list of metrics which should be calculated. Currently supporting: 'mae' (mean absolute error), 'rmse' (root mean squared error), and 'wis' (weighted interval score)

horizons

vector of horizons for which the metrics be calculated. Default is to calculate metrics only for horizon 0 (the nowcast).

Examples

#These examples require the `scoringutils` function
if (requireNamespace("scoringutils", quietly = TRUE)){

# Load the data
data(denguedat)

# In this example we will test two models
now    <- as.Date("1990-10-01")
ncast1 <- nowcast(denguedat, "onset_week", "report_week", now = now,
  method = "optimization", seed = 2495624, iter = 10)

ncast2 <- nowcast(denguedat, "onset_week", "report_week", now = now,
  method = "optimization", seed = 2495624, iter = 10, dist = "Normal")

# Run a backtest for each of the models
btest1 <- backtest(ncast1, dates_to_test = as.Date("1990-06-11"), model_name = "Classic")
btest2 <- backtest(ncast2, dates_to_test = as.Date("1990-06-11"), model_name = "Normal")

# Compare the models to select the best model
backtest_metrics(btest1, btest2)
}
#> # A tibble: 2 × 14
#>   horizon Strata_unified model   now          mse se_point   wis overprediction
#>     <dbl> <chr>          <chr>   <date>     <dbl>    <dbl> <dbl>          <dbl>
#> 1       0 No strata      Classic 1990-06-11  24.1     582. 13.0           11   
#> 2       0 No strata      Normal  1990-06-11  13.3     177.  5.61           3.83
#> # ℹ 6 more variables: underprediction <dbl>, dispersion <dbl>, bias <dbl>,
#> #   interval_coverage_50 <dbl>, interval_coverage_90 <dbl>, ae_median <dbl>