Build the default prior bundle for an RTMB nowcast model

Build the default prior bundle for an RTMB nowcast model

Usage

default_priors(mod, data = NULL, ...)

Arguments

mod

A model() object.

data

Optional prepared-data list from prepare_data() (used for the data-informed location/scale defaults). May also be a bare list with an m matrix.

...

Per-key overrides (e.g. phi = lognormal_prior(log(20), 0.5), delay_mu = normal_prior(log(5), 0.3)).

Value

A named list of prior specs.

Details

The default prior for each parameter is documented on the component constructor: see the Default priors section of epidemic_process (HSGP / AR(1) / SIR), delay_process (LogNormal / Gamma / GenGamma / Dirichlet) and likelihood (the NB overdispersion phi). To see what a set of priors implies for the epidemic curve before fitting, use nowcast(prior_only = TRUE).

See also