Automatically infer which value is `units`.

Function returns whether data is daily, weekly, monthly or yearly by `date_column`.

Usage

infer_units(.disease_data, units, date_column)

Arguments

.disease_data

A time series of reporting data in aggregated line list format such that each row has a column for onset date, report date, and

units

Time scale of reporting. Options: "1 day", "1 week".

date_column

Name of a column of `.disease_data` that contains the dates.

Value

Whether the data's units are `days` or `weeks`