Nowcasting with the `rstan` engine

Nowcasting with the `rstan` engine

Usage

nowcast.rstan(
  .disease_data,
  onset_date,
  report_date,
  strata = NULL,
  dist = c("NegativeBinomial", "Poisson"),
  prior_only = FALSE,
  control = control_default(),
  refresh = 250,
  mu_degree = 2,
  nu_degree = 1,
  mu_is_constant = FALSE,
  nu_is_constant = TRUE,
  mu_error_prior = "normal",
  nu_error_prior = "normal",
  mu_param_1 = 0,
  mu_param_2 = 1,
  nu_param_1 = 0,
  nu_param_2 = 1,
  mu_0_prior = "normal",
  nu_0_prior = "normal",
  mu_0_param_1 = log(mean(.disease_data$n, na.rm = T)),
  mu_0_param_2 = 0.01,
  nu_0_param_1 = 0,
  nu_0_param_2 = 0.01,
  r_prior = "normal",
  r_param_1 = 0,
  r_param_2 = 1,
  ...
)

Arguments

.disease_data

A time series of reporting data in aggregated line list format such that each row has a column for onset date, report date, and

onset_date

In quotations, the name of the column of datatype Date designating the date of case onset. e.g. "onset_week"

report_date

In quotations, the name of the column of datatype Date designating the date of case report. e.g. "report_week"

strata

Character vector of names of the strata included in the data.

dist

Distribution. Either "NegativeBinomial" or "Poisson"

prior_only

Boolean variable indicating whether to compute only the prior distribution

control

Control parameter for [rstan::sampling()]

refresh

Refresh parameter for [rstan::sampling()]

mu_degree

Integer. Degree of the epidemic trend. Default is 2.

nu_degree

Integer. Degree of the delay trend. Default is 1.

mu_is_constant

Logical. Indicates whether the epidemic trend is constant. Default is FALSE.

nu_is_constant

Logical. Indicates whether the delay trend is constant. Default is TRUE.

mu_error_prior

Character. Prior for the epidemic trend error. Default is "normal".

nu_error_prior

Character. Prior for the delay trend error. Default is "normal".

mu_param_1

Numeric. First parameter for the epidemic trend error. Default is 0.0.

mu_param_2

Numeric. Second parameter for the epidemic trend error. Default is 1.0.

nu_param_1

Numeric. First parameter for the delay trend error. Default is 0.0.

nu_param_2

Numeric. Second parameter for the delay trend error. Default is 1.0.

mu_0_prior

Character. Prior for the initial epidemic trend distribution. Default is "normal".

nu_0_prior

Character. Prior for the initial delay trend distribution. Default is "normal".

mu_0_param_1

Numeric. First parameter for the initial epidemic trend distribution. Default is 0.0.

mu_0_param_2

Numeric. Second parameter for the initial epidemic trend distribution. Default is 1.0.

nu_0_param_1

Numeric. First parameter for the initial delay trend distribution. Default is 0.0.

nu_0_param_2

Numeric. Second parameter for the initial delay trend distribution. Default is 1.0.

r_prior

Character. Prior for the negative binomial precision parameter. Default is "normal".

r_param_1

Numeric. First parameter for the dispersion prior if negative binomial. Default is 0.0.

r_param_2

Numeric. Second parameter for the dispersion prior if negative binomial. Default is 1.0.

...

Additional arguments to pass to [rstan::sample()]