Get nowcast priors not controlling by distribution

Function to return the default values for the nowcast() function as a list.

Usage

set_priors(
  mu_intercept_param_1 = 0,
  mu_intercept_param_2 = 1,
  mu_0_param_1 = 0,
  mu_0_param_2 = 1,
  nu_intercept_param_1 = 0,
  nu_intercept_param_2 = 1,
  nu_0_param_1 = 0,
  nu_0_param_2 = 1,
  c_0_param_1 = 0,
  c_0_param_2 = 1,
  ctilde_0_param_1 = 0,
  ctilde_0_param_2 = 1,
  sd_mu_param_1 = 0,
  sd_mu_param_2 = 1,
  sd_nu_param_1 = 0,
  sd_nu_param_2 = 1,
  sd_c_param_1 = 0,
  sd_c_param_2 = 1,
  sd_ctilde_param_1 = 0,
  sd_ctilde_param_2 = 1,
  sd_m_param_1 = 0,
  sd_m_param_2 = 1,
  sd_dow_epi_param_1 = 0,
  sd_dow_epi_param_2 = 1,
  sd_wkend_epi_param_1 = 0,
  sd_wkend_epi_param_2 = 1,
  sd_dom_epi_param_1 = 0,
  sd_dom_epi_param_2 = 1,
  sd_month_epi_param_1 = 0,
  sd_month_epi_param_2 = 1,
  sd_week_epi_param_1 = 0,
  sd_week_epi_param_2 = 1,
  sd_holidays_epi_param_1 = 0,
  sd_holidays_epi_param_2 = 1,
  dof = 7,
  control_k_transform = 2,
  control_c_transform = 0.5
)

Arguments

mu_intercept_param_1

Real. Mean of the epidemic intercept.

mu_intercept_param_2

Positive real. Variance of the epidemic intercept.

mu_0_param_1

Real. Mean of the initial value of the epidemic trend.

mu_0_param_2

Positive Real. Variance of the initial value of the epidemic trend.

nu_intercept_param_1

Real. Mean of the delay intercept.

nu_intercept_param_2

Positive real. Variance of the delay intercept.

nu_0_param_1

Real. Mean of the initial value of the disease trend.

nu_0_param_2

Positive Real. Variance of the initial value of the disease trend.

c_0_param_1

Initial value for cycle parameter's mean

c_0_param_2

Initial value for cycle parameter's sd

ctilde_0_param_1

Initial value for cycle parameter's mean (latent cycle)

ctilde_0_param_2

Initial value for cycle parameter's sd (latent cycle)

sd_mu_param_1

Positive real. Initial value for the mean of the standard deviation of the epidemic trend.

sd_mu_param_2

Positive real. Initial value for the variance of the standard deviation of the epidemic trend.

sd_nu_param_1

Positive real. Initial value for the mean of the standard deviation of the disease trend.

sd_nu_param_2

Positive real. Initial value for the variance of the standard deviation of the disease trend.

sd_c_param_1

Mean value for the standard deviation's of c_0

sd_c_param_2

Standard deviation prior value for the standard deviation's of c_0

sd_ctilde_param_1

Mean value for the standard deviation's of c_0 tilde

sd_ctilde_param_2

Standard deviation prior value for the standard deviation's of c_0 tilde

sd_m_param_1

Positive real. Initial value for the mean of the standard deviation of the observed cases (not considered if distribution is Poisson)

sd_m_param_2

Positive real. Initial value for the variance of the standard deviation of the observed cases (not considered if distribution is Poisson)

sd_dow_epi_param_1

Standard deviation's mean for day of the week effect

sd_dow_epi_param_2

Standard deviation's sd for day of the week effect

sd_wkend_epi_param_1

Standard deviation's mean for weekend effect

sd_wkend_epi_param_2

Standard deviation's sd for weekend effect

sd_dom_epi_param_1

Standard deviation's mean for day of the month effect

sd_dom_epi_param_2

Standard deviation's sd for day of the month effect

sd_month_epi_param_1

Standard deviation's mean for month effect

sd_month_epi_param_2

Standard deviation's sd for month effect

sd_week_epi_param_1

Standard deviation's mean for epiweek effect

sd_week_epi_param_2

Standard deviation's sd for epiweek effect

sd_holidays_epi_param_1

Standard deviation's mean for holiday effect

sd_holidays_epi_param_2

Standard deviation's sd for holiday effect

dof

Degrees of freedom if the distribution used in nowcast() is Student (default = 7).

control_k_transform

Parameter for the dhyperbolic or softplus link functions.

control_c_transform

Parameter for the dhyperbolic link function.

Value

A list with all the priors for the nowcast() function.

Examples

#Get the default priors
set_priors()
#> $mu_intercept_param_1
#> [1] 0
#> 
#> $mu_intercept_param_2
#> [1] 1
#> 
#> $mu_0_param_1
#> [1] 0
#> 
#> $mu_0_param_2
#> [1] 1
#> 
#> $nu_intercept_param_1
#> [1] 0
#> 
#> $nu_intercept_param_2
#> [1] 1
#> 
#> $nu_0_param_1
#> [1] 0
#> 
#> $nu_0_param_2
#> [1] 1
#> 
#> $c_0_param_1
#> [1] 0
#> 
#> $c_0_param_2
#> [1] 1
#> 
#> $ctilde_0_param_1
#> [1] 0
#> 
#> $ctilde_0_param_2
#> [1] 1
#> 
#> $sd_mu_param_1
#> [1] 0
#> 
#> $sd_mu_param_2
#> [1] 1
#> 
#> $sd_nu_param_1
#> [1] 0
#> 
#> $sd_nu_param_2
#> [1] 1
#> 
#> $sd_c_param_1
#> [1] 0
#> 
#> $sd_c_param_2
#> [1] 1
#> 
#> $sd_ctilde_param_1
#> [1] 0
#> 
#> $sd_ctilde_param_2
#> [1] 1
#> 
#> $sd_m_param_1
#> [1] 0
#> 
#> $sd_m_param_2
#> [1] 1
#> 
#> $sd_dow_epi_param_1
#> [1] 0
#> 
#> $sd_dow_epi_param_2
#> [1] 1
#> 
#> $sd_wkend_epi_param_1
#> [1] 0
#> 
#> $sd_wkend_epi_param_2
#> [1] 1
#> 
#> $sd_dom_epi_param_1
#> [1] 0
#> 
#> $sd_dom_epi_param_2
#> [1] 1
#> 
#> $sd_month_epi_param_1
#> [1] 0
#> 
#> $sd_month_epi_param_2
#> [1] 1
#> 
#> $sd_week_epi_param_1
#> [1] 0
#> 
#> $sd_week_epi_param_2
#> [1] 1
#> 
#> $sd_holidays_epi_param_1
#> [1] 0
#> 
#> $sd_holidays_epi_param_2
#> [1] 1
#> 
#> $dof
#> [1] 7
#> 
#> $control_k_transform
#> [1] 2
#> 
#> $control_c_transform
#> [1] 0.5
#> 

#Change the priors
set_priors(mu_intercept_param_1 = 2)
#> $mu_intercept_param_1
#> [1] 2
#> 
#> $mu_intercept_param_2
#> [1] 1
#> 
#> $mu_0_param_1
#> [1] 0
#> 
#> $mu_0_param_2
#> [1] 1
#> 
#> $nu_intercept_param_1
#> [1] 0
#> 
#> $nu_intercept_param_2
#> [1] 1
#> 
#> $nu_0_param_1
#> [1] 0
#> 
#> $nu_0_param_2
#> [1] 1
#> 
#> $c_0_param_1
#> [1] 0
#> 
#> $c_0_param_2
#> [1] 1
#> 
#> $ctilde_0_param_1
#> [1] 0
#> 
#> $ctilde_0_param_2
#> [1] 1
#> 
#> $sd_mu_param_1
#> [1] 0
#> 
#> $sd_mu_param_2
#> [1] 1
#> 
#> $sd_nu_param_1
#> [1] 0
#> 
#> $sd_nu_param_2
#> [1] 1
#> 
#> $sd_c_param_1
#> [1] 0
#> 
#> $sd_c_param_2
#> [1] 1
#> 
#> $sd_ctilde_param_1
#> [1] 0
#> 
#> $sd_ctilde_param_2
#> [1] 1
#> 
#> $sd_m_param_1
#> [1] 0
#> 
#> $sd_m_param_2
#> [1] 1
#> 
#> $sd_dow_epi_param_1
#> [1] 0
#> 
#> $sd_dow_epi_param_2
#> [1] 1
#> 
#> $sd_wkend_epi_param_1
#> [1] 0
#> 
#> $sd_wkend_epi_param_2
#> [1] 1
#> 
#> $sd_dom_epi_param_1
#> [1] 0
#> 
#> $sd_dom_epi_param_2
#> [1] 1
#> 
#> $sd_month_epi_param_1
#> [1] 0
#> 
#> $sd_month_epi_param_2
#> [1] 1
#> 
#> $sd_week_epi_param_1
#> [1] 0
#> 
#> $sd_week_epi_param_2
#> [1] 1
#> 
#> $sd_holidays_epi_param_1
#> [1] 0
#> 
#> $sd_holidays_epi_param_2
#> [1] 1
#> 
#> $dof
#> [1] 7
#> 
#> $control_k_transform
#> [1] 2
#> 
#> $control_c_transform
#> [1] 0.5
#>