Nowcasting

Function that uses the [rstan::sampling()] engine to generate nowcasts.

Usage

nowcast(
  .disease_data,
  onset_date,
  report_date,
  strata = NULL,
  dist = c("NegativeBinomial", "Poisson"),
  now = NULL,
  units = NULL,
  max_delay = Inf,
  prior_only = FALSE,
  proportion_reported = 1,
  refresh = 250 * interactive(),
  control = control_default(),
  method = c("sampling", "variational", "optimization"),
  priors = set_priors(),
  ...
)

Arguments

.disease_data

A time series of reporting data in aggregated line list format such that each row has a column for onset date, report date, and

onset_date

In quotations, the name of the column of datatype Date designating the date of case onset. e.g. "onset_week"

report_date

In quotations, the name of the column of datatype Date designating the date of case report. e.g. "report_week"

strata

Character vector of names of the strata included in the data.

dist

Distribution. Either "NegativeBinomial" or "Poisson"

now

An object of datatype Date indicating the date at which to perform the nowcast.

units

Time scale of reporting. Options: "1 day", "1 week".

max_delay

Maximum possible delay observed or considered for estimation of the delay distribution (numeric). Default: `Inf`

prior_only

Boolean variable indicating whether to compute only the prior distribution

proportion_reported

A decimal greater than 0 and less than or equal to 1 representing the proportion of all cases expected to be reported. Default: 1, e.g. 100 percent of all cases will eventually be reported. For asymptomatic diseases where not all cases will ever be reported, or for outbreaks in which severe under-reporting is expected, change this to less than 1.

refresh

Refresh parameter for [rstan::sampling()]

control

Control parameter for [rstan::sampling()]

method

Fitting method either `sampling` (recommended for inference), `variational` (recommended for testing) or `optimization`. The `sampling` method calls [rstan::sampling()] while the `variational` calls [rstan::vb()] and `optimization` calls [rstan::optimizing()]

priors

A list of all of the nowcast priors. You can use [set_priors()] to change the priors of the function (see details)

...

Additional arguments to pass to [rstan::sampling()]

Examples

# Load the data
data(denguedat)

# Create a fake disease process
sims <- simulate_process_for_testing()

# Run a nowcast with very few iterations
# change to 4 chains and 2000 iter when doing inference
nowcast(sims, "onset_date", "report_date", method = "optimization", seed = 2524)
#> ℹ Computing a nowcast for 2024-09-19 per "days"
#> ℹ Assuming data is count-data where counts are in column `n`. To change this set `data_type = "linelist"`
#> $data
#> $data$stan_data
#> $data$stan_data$num_steps
#> [1] 10
#> 
#> $data$stan_data$num_delays
#> [1] 8
#> 
#> $data$stan_data$num_strata
#> [1] 1
#> 
#> $data$stan_data$n_rows
#> [1] 52
#> 
#> $data$stan_data$N_cases
#>          n .tval .delay .strata
#>  [1,]    1     1      1       1
#>  [2,]    0     1      2       1
#>  [3,]    0     1      3       1
#>  [4,]    0     1      4       1
#>  [5,]    1     1      5       1
#>  [6,]    1     1      6       1
#>  [7,]    1     1      7       1
#>  [8,]    2     1      8       1
#>  [9,]    0     2      1       1
#> [10,]    0     2      2       1
#> [11,]    3     2      3       1
#> [12,]    0     2      4       1
#> [13,]    0     2      5       1
#> [14,]    3     2      6       1
#> [15,]    4     2      7       1
#> [16,]    0     2      8       1
#> [17,]    0     3      1       1
#> [18,]    7     3      2       1
#> [19,]    5     3      3       1
#> [20,]    5     3      4       1
#> [21,]   30     3      5       1
#> [22,]    0     3      6       1
#> [23,]    0     3      7       1
#> [24,]    0     3      8       1
#> [25,]    2     4      1       1
#> [26,]    8     4      2       1
#> [27,]  146     4      3       1
#> [28,]    1     4      4       1
#> [29,]   98     4      5       1
#> [30,]    2     4      6       1
#> [31,]    0     4      7       1
#> [32,]    0     5      1       1
#> [33,]    0     5      2       1
#> [34,]   58     5      3       1
#> [35,]    5     5      4       1
#> [36,] 3458     5      5       1
#> [37,]    3     5      6       1
#> [38,]    0     6      1       1
#> [39,]    2     6      2       1
#> [40,]  516     6      3       1
#> [41,]   19     6      4       1
#> [42,] 9619     6      5       1
#> [43,]    0     7      1       1
#> [44,]    0     7      2       1
#> [45,]   26     7      3       1
#> [46,]   14     7      4       1
#> [47,]    6     8      1       1
#> [48,]    1     8      2       1
#> [49,] 2213     8      3       1
#> [50,]    9     9      1       1
#> [51,]    0     9      2       1
#> [52,]  107    10      1       1
#> 
#> $data$stan_data$prior_only
#> [1] 0
#> 
#> $data$stan_data$mu_degree
#> [1] 1
#> 
#> $data$stan_data$nu_degree
#> [1] 1
#> 
#> $data$stan_data$mu_is_constant
#> [1] FALSE
#> 
#> $data$stan_data$nu_is_constant
#> [1] TRUE
#> 
#> $data$stan_data$is_negative_binomial
#> [1] 1
#> 
#> $data$stan_data$mu_sd_prior
#> [1] 1
#> 
#> $data$stan_data$nu_sd_prior
#> [1] 1
#> 
#> $data$stan_data$r_prior
#> [1] 1
#> 
#> $data$stan_data$r_param_1
#> [1] 0
#> 
#> $data$stan_data$r_param_2
#> [1] 1
#> 
#> $data$stan_data$mu_sd_param_1
#> [1] 0
#> 
#> $data$stan_data$mu_sd_param_2
#> [1] 0.1
#> 
#> $data$stan_data$nu_sd_param_1
#> [1] 0
#> 
#> $data$stan_data$nu_sd_param_2
#> [1] 0.1
#> 
#> $data$stan_data$mu_0_mean_param_1
#> [1] 1.735997
#> 
#> $data$stan_data$mu_0_mean_param_2
#> [1] 0.01
#> 
#> $data$stan_data$mu_0_sd_param_1
#> [1] 2.25634
#> 
#> $data$stan_data$mu_0_sd_param_2
#> [1] 0.01
#> 
#> $data$stan_data$nu_0_mean_param_1
#> [1] 0
#> 
#> $data$stan_data$nu_0_mean_param_2
#> [1] 0.01
#> 
#> $data$stan_data$nu_0_sd_param_1
#> [1] 0
#> 
#> $data$stan_data$nu_0_sd_param_2
#> [1] 0.01
#> 
#> $data$stan_data$mu_0_mean_hyperprior
#> [1] 1
#> 
#> $data$stan_data$nu_0_mean_hyperprior
#> [1] 1
#> 
#> $data$stan_data$mu_0_sd_hyperprior
#> [1] 1
#> 
#> $data$stan_data$nu_0_sd_hyperprior
#> [1] 1
#> 
#> 
#> $data$preprocessed_data
#> # A tibble: 52 × 5
#>    onset_date .delay report_date     n .tval
#>    <date>      <dbl> <date>      <dbl> <dbl>
#>  1 2024-09-10      0 2024-09-10      1     1
#>  2 2024-09-10      1 2024-09-11      0     1
#>  3 2024-09-10      2 2024-09-12      0     1
#>  4 2024-09-10      3 2024-09-13      0     1
#>  5 2024-09-10      4 2024-09-14      1     1
#>  6 2024-09-10      5 2024-09-15      1     1
#>  7 2024-09-10      6 2024-09-16      1     1
#>  8 2024-09-10      7 2024-09-17      2     1
#>  9 2024-09-11      0 2024-09-11      0     2
#> 10 2024-09-11      1 2024-09-12      0     2
#> # ℹ 42 more rows
#> 
#> $data$call_parameters
#> $data$call_parameters$onset_date
#> [1] "onset_date"
#> 
#> $data$call_parameters$report_date
#> [1] "report_date"
#> 
#> $data$call_parameters$strata
#> NULL
#> 
#> $data$call_parameters$now
#> [1] "2024-09-19"
#> 
#> $data$call_parameters$units
#> [1] "days"
#> 
#> $data$call_parameters$max_delay
#> [1] Inf
#> 
#> 
#> 
#> $dict
#> $dict$strata_dict
#> # A tibble: 1 × 2
#>   .strata_unified .strata
#>   <chr>             <dbl>
#> 1 No strata             1
#> 
#> 
#> $generated_quantities
#> Inference for Stan model: generated_quantities.
#> 1 chains, each with iter=1000; warmup=0; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=1000.
#> 
#>                         mean se_mean      sd     2.5%      25%     50%      75%
#> N_mat_predict[1,1]      1.00     NaN    0.00     1.00     1.00     1.0     1.00
#> N_mat_predict[1,2]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[1,3]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[1,4]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[1,5]      1.00     NaN    0.00     1.00     1.00     1.0     1.00
#> N_mat_predict[1,6]      1.00     NaN    0.00     1.00     1.00     1.0     1.00
#> N_mat_predict[1,7]      1.00     NaN    0.00     1.00     1.00     1.0     1.00
#> N_mat_predict[1,8]      2.00     NaN    0.00     2.00     2.00     2.0     2.00
#> N_mat_predict[2,1]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[2,2]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[2,3]      3.00     NaN    0.00     3.00     3.00     3.0     3.00
#> N_mat_predict[2,4]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[2,5]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[2,6]      3.00     NaN    0.00     3.00     3.00     3.0     3.00
#> N_mat_predict[2,7]      4.00     NaN    0.00     4.00     4.00     4.0     4.00
#> N_mat_predict[2,8]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[3,1]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[3,2]      7.00     NaN    0.00     7.00     7.00     7.0     7.00
#> N_mat_predict[3,3]      5.00     NaN    0.00     5.00     5.00     5.0     5.00
#> N_mat_predict[3,4]      5.00     NaN    0.00     5.00     5.00     5.0     5.00
#> N_mat_predict[3,5]     30.00     NaN    0.00    30.00    30.00    30.0    30.00
#> N_mat_predict[3,6]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[3,7]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[3,8]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[4,1]      2.00     NaN    0.00     2.00     2.00     2.0     2.00
#> N_mat_predict[4,2]      8.00     NaN    0.00     8.00     8.00     8.0     8.00
#> N_mat_predict[4,3]    146.00     NaN    0.00   146.00   146.00   146.0   146.00
#> N_mat_predict[4,4]      1.00     NaN    0.00     1.00     1.00     1.0     1.00
#> N_mat_predict[4,5]     98.00     NaN    0.00    98.00    98.00    98.0    98.00
#> N_mat_predict[4,6]      2.00     NaN    0.00     2.00     2.00     2.0     2.00
#> N_mat_predict[4,7]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[4,8]      0.06    0.01    0.25     0.00     0.00     0.0     0.00
#> N_mat_predict[5,1]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[5,2]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[5,3]     58.00     NaN    0.00    58.00    58.00    58.0    58.00
#> N_mat_predict[5,4]      5.00     NaN    0.00     5.00     5.00     5.0     5.00
#> N_mat_predict[5,5]   3458.00     NaN    0.00  3458.00  3458.00  3458.0  3458.00
#> N_mat_predict[5,6]      3.00     NaN    0.00     3.00     3.00     3.0     3.00
#> N_mat_predict[5,7]      0.07    0.01    0.25     0.00     0.00     0.0     0.00
#> N_mat_predict[5,8]      0.06    0.01    0.23     0.00     0.00     0.0     0.00
#> N_mat_predict[6,1]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[6,2]      2.00     NaN    0.00     2.00     2.00     2.0     2.00
#> N_mat_predict[6,3]    516.00     NaN    0.00   516.00   516.00   516.0   516.00
#> N_mat_predict[6,4]     19.00     NaN    0.00    19.00    19.00    19.0    19.00
#> N_mat_predict[6,5]   9619.00     NaN    0.00  9619.00  9619.00  9619.0  9619.00
#> N_mat_predict[6,6]      2.80    0.07    2.29     0.00     1.00     2.0     4.00
#> N_mat_predict[6,7]      0.07    0.01    0.27     0.00     0.00     0.0     0.00
#> N_mat_predict[6,8]      0.04    0.01    0.21     0.00     0.00     0.0     0.00
#> N_mat_predict[7,1]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[7,2]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[7,3]     26.00     NaN    0.00    26.00    26.00    26.0    26.00
#> N_mat_predict[7,4]     14.00     NaN    0.00    14.00    14.00    14.0    14.00
#> N_mat_predict[7,5]   9514.05  203.32 5347.75  1837.18  5868.00  8444.0 12272.50
#> N_mat_predict[7,6]      3.02    0.08    2.54     0.00     1.00     2.0     4.00
#> N_mat_predict[7,7]      0.07    0.01    0.26     0.00     0.00     0.0     0.00
#> N_mat_predict[7,8]      0.05    0.01    0.21     0.00     0.00     0.0     0.00
#> N_mat_predict[8,1]      6.00     NaN    0.00     6.00     6.00     6.0     6.00
#> N_mat_predict[8,2]      1.00     NaN    0.00     1.00     1.00     1.0     1.00
#> N_mat_predict[8,3]   2213.00     NaN    0.00  2213.00  2213.00  2213.0  2213.00
#> N_mat_predict[8,4]     14.24    0.28    8.75     2.00     8.00    12.0    19.00
#> N_mat_predict[8,5]   9556.26  177.55 5419.61  2149.88  5733.25  8567.5 12235.50
#> N_mat_predict[8,6]      2.96    0.07    2.33     0.00     1.00     2.0     4.00
#> N_mat_predict[8,7]      0.07    0.01    0.27     0.00     0.00     0.0     0.00
#> N_mat_predict[8,8]      0.06    0.01    0.25     0.00     0.00     0.0     0.00
#> N_mat_predict[9,1]      9.00     NaN    0.00     9.00     9.00     9.0     9.00
#> N_mat_predict[9,2]      0.00     NaN    0.00     0.00     0.00     0.0     0.00
#> N_mat_predict[9,3]   2072.27   34.08 1143.73   449.70  1227.00  1858.0  2676.75
#> N_mat_predict[9,4]     14.13    0.28    8.66     2.00     8.00    13.0    19.00
#> N_mat_predict[9,5]   9259.91  163.06 5342.29  2103.78  5648.75  8054.0 11775.75
#> N_mat_predict[9,6]      3.00    0.08    2.49     0.00     1.00     2.0     4.00
#> N_mat_predict[9,7]      0.06    0.01    0.26     0.00     0.00     0.0     0.00
#> N_mat_predict[9,8]      0.05    0.01    0.23     0.00     0.00     0.0     0.00
#> N_mat_predict[10,1]   107.00     NaN    0.00   107.00   107.00   107.0   107.00
#> N_mat_predict[10,2]     0.13    0.01    0.35     0.00     0.00     0.0     0.00
#> N_mat_predict[10,3]  2049.14   37.42 1176.56   436.75  1185.00  1835.5  2677.50
#> N_mat_predict[10,4]    14.16    0.27    8.78     2.00     8.00    13.0    19.00
#> N_mat_predict[10,5]  9432.07  170.05 5216.34  2246.52  5641.50  8328.0 12082.00
#> N_mat_predict[10,6]     2.95    0.08    2.43     0.00     1.00     2.0     4.00
#> N_mat_predict[10,7]     0.06    0.01    0.26     0.00     0.00     0.0     0.00
#> N_mat_predict[10,8]     0.05    0.01    0.22     0.00     0.00     0.0     0.00
#> N_predict[1,1]          6.00     NaN    0.00     6.00     6.00     6.0     6.00
#> N_predict[2,1]         10.00     NaN    0.00    10.00    10.00    10.0    10.00
#> N_predict[3,1]         47.00     NaN    0.00    47.00    47.00    47.0    47.00
#> N_predict[4,1]        257.06    0.01    0.25   257.00   257.00   257.0   257.00
#> N_predict[5,1]       3524.12    0.01    0.35  3524.00  3524.00  3524.0  3524.00
#> N_predict[6,1]      10158.92    0.07    2.31 10156.00 10157.00 10158.0 10160.00
#> N_predict[7,1]       9557.18  203.35 5347.85  1882.08  5912.00  8485.5 12312.75
#> N_predict[8,1]      11793.59  177.51 5419.17  4384.25  7966.25 10796.5 14470.75
#> N_predict[9,1]      11358.43  168.50 5436.56  3758.88  7674.50 10190.5 14192.00
#> N_predict[10,1]     11605.56  171.80 5332.28  4091.95  7821.00 10671.5 14433.25
#>                        97.5% n_eff Rhat
#> N_mat_predict[1,1]      1.00   NaN  NaN
#> N_mat_predict[1,2]      0.00   NaN  NaN
#> N_mat_predict[1,3]      0.00   NaN  NaN
#> N_mat_predict[1,4]      0.00   NaN  NaN
#> N_mat_predict[1,5]      1.00   NaN  NaN
#> N_mat_predict[1,6]      1.00   NaN  NaN
#> N_mat_predict[1,7]      1.00   NaN  NaN
#> N_mat_predict[1,8]      2.00   NaN  NaN
#> N_mat_predict[2,1]      0.00   NaN  NaN
#> N_mat_predict[2,2]      0.00   NaN  NaN
#> N_mat_predict[2,3]      3.00   NaN  NaN
#> N_mat_predict[2,4]      0.00   NaN  NaN
#> N_mat_predict[2,5]      0.00   NaN  NaN
#> N_mat_predict[2,6]      3.00   NaN  NaN
#> N_mat_predict[2,7]      4.00   NaN  NaN
#> N_mat_predict[2,8]      0.00   NaN  NaN
#> N_mat_predict[3,1]      0.00   NaN  NaN
#> N_mat_predict[3,2]      7.00   NaN  NaN
#> N_mat_predict[3,3]      5.00   NaN  NaN
#> N_mat_predict[3,4]      5.00   NaN  NaN
#> N_mat_predict[3,5]     30.00   NaN  NaN
#> N_mat_predict[3,6]      0.00   NaN  NaN
#> N_mat_predict[3,7]      0.00   NaN  NaN
#> N_mat_predict[3,8]      0.00   NaN  NaN
#> N_mat_predict[4,1]      2.00   NaN  NaN
#> N_mat_predict[4,2]      8.00   NaN  NaN
#> N_mat_predict[4,3]    146.00   NaN  NaN
#> N_mat_predict[4,4]      1.00   NaN  NaN
#> N_mat_predict[4,5]     98.00   NaN  NaN
#> N_mat_predict[4,6]      2.00   NaN  NaN
#> N_mat_predict[4,7]      0.00   NaN  NaN
#> N_mat_predict[4,8]      1.00   997 1.00
#> N_mat_predict[5,1]      0.00   NaN  NaN
#> N_mat_predict[5,2]      0.00   NaN  NaN
#> N_mat_predict[5,3]     58.00   NaN  NaN
#> N_mat_predict[5,4]      5.00   NaN  NaN
#> N_mat_predict[5,5]   3458.00   NaN  NaN
#> N_mat_predict[5,6]      3.00   NaN  NaN
#> N_mat_predict[5,7]      1.00   798 1.00
#> N_mat_predict[5,8]      1.00   901 1.00
#> N_mat_predict[6,1]      0.00   NaN  NaN
#> N_mat_predict[6,2]      2.00   NaN  NaN
#> N_mat_predict[6,3]    516.00   NaN  NaN
#> N_mat_predict[6,4]     19.00   NaN  NaN
#> N_mat_predict[6,5]   9619.00   NaN  NaN
#> N_mat_predict[6,6]      9.00  1009 1.00
#> N_mat_predict[6,7]      1.00  1016 1.00
#> N_mat_predict[6,8]      1.00   972 1.00
#> N_mat_predict[7,1]      0.00   NaN  NaN
#> N_mat_predict[7,2]      0.00   NaN  NaN
#> N_mat_predict[7,3]     26.00   NaN  NaN
#> N_mat_predict[7,4]     14.00   NaN  NaN
#> N_mat_predict[7,5]  23161.38   692 1.00
#> N_mat_predict[7,6]      9.00   923 1.00
#> N_mat_predict[7,7]      1.00   850 1.00
#> N_mat_predict[7,8]      1.00   879 1.00
#> N_mat_predict[8,1]      6.00   NaN  NaN
#> N_mat_predict[8,2]      1.00   NaN  NaN
#> N_mat_predict[8,3]   2213.00   NaN  NaN
#> N_mat_predict[8,4]     36.00   952 1.00
#> N_mat_predict[8,5]  22532.32   932 1.00
#> N_mat_predict[8,6]      9.00   990 1.01
#> N_mat_predict[8,7]      1.00   872 1.00
#> N_mat_predict[8,8]      1.00   883 1.01
#> N_mat_predict[9,1]      9.00   NaN  NaN
#> N_mat_predict[9,2]      0.00   NaN  NaN
#> N_mat_predict[9,3]   4931.20  1126 1.00
#> N_mat_predict[9,4]     36.00   930 1.00
#> N_mat_predict[9,5]  22024.22  1073 1.00
#> N_mat_predict[9,6]      9.00  1015 1.00
#> N_mat_predict[9,7]      1.00   825 1.00
#> N_mat_predict[9,8]      1.00  1004 1.00
#> N_mat_predict[10,1]   107.00   NaN  NaN
#> N_mat_predict[10,2]     1.00   819 1.00
#> N_mat_predict[10,3]  4901.08   989 1.00
#> N_mat_predict[10,4]    35.00  1058 1.00
#> N_mat_predict[10,5] 22063.52   941 1.01
#> N_mat_predict[10,6]     9.00   930 1.00
#> N_mat_predict[10,7]     1.00   842 1.00
#> N_mat_predict[10,8]     1.00   990 1.00
#> N_predict[1,1]          6.00   NaN  NaN
#> N_predict[2,1]         10.00   NaN  NaN
#> N_predict[3,1]         47.00   NaN  NaN
#> N_predict[4,1]        258.00   997 1.00
#> N_predict[5,1]       3525.00   853 1.00
#> N_predict[6,1]      10165.00   968 1.00
#> N_predict[7,1]      23204.40   692 1.00
#> N_predict[8,1]      24772.32   932 1.00
#> N_predict[9,1]      24935.12  1041 1.00
#> N_predict[10,1]     24642.45   963 1.00
#> 
#> Samples were drawn using  at Fri Sep 20 06:48:43 2024.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).
#> 
#> $model
#> $model$par
#>    mu_0_centered[1,1]    mu_0_centered[2,1]    mu_0_centered[3,1] 
#>          5.653347e-02         -4.022521e-01         -2.944901e-01 
#>    mu_0_centered[4,1]    mu_0_centered[5,1]    mu_0_centered[6,1] 
#>         -4.061945e-01          1.018126e-01          1.759186e-01 
#>    mu_0_centered[7,1]    mu_0_centered[8,1]    nu_0_centered[1,1] 
#>          1.863656e-01          4.481930e-01          1.028474e-02 
#>    nu_0_centered[2,1]    nu_0_centered[3,1]    nu_0_centered[4,1] 
#>         -7.051117e-02         -5.152598e-02         -7.107033e-02 
#>    nu_0_centered[5,1]    nu_0_centered[6,1]    nu_0_centered[7,1] 
#>          1.834992e-02          3.094108e-02          3.270211e-02 
#>    nu_0_centered[8,1]             mu_0_mean             nu_0_mean 
#>          7.863037e-02         -1.620234e-01         -1.623436e-01 
#>               mu_0_sd               nu_0_sd                 mu_sd 
#>          8.342856e-01          1.465672e-01          3.626036e+00 
#>                 nu_sd xi_mu_centered[1,1,1] xi_mu_centered[2,1,1] 
#>          4.294389e-02         -4.594383e-01          3.870599e-02 
#> xi_mu_centered[3,1,1] xi_mu_centered[4,1,1] xi_mu_centered[5,1,1] 
#>          6.080909e-01         -6.374715e-01         -1.236055e-01 
#> xi_mu_centered[6,1,1] xi_mu_centered[7,1,1] xi_mu_centered[8,1,1] 
#>          2.112292e-01          9.059053e-01          1.546413e-01 
#> xi_mu_centered[9,1,1] xi_mu_centered[1,2,1] xi_mu_centered[2,2,1] 
#>          6.513463e-01         -1.486009e-01          8.430429e-01 
#> xi_mu_centered[3,2,1] xi_mu_centered[4,2,1] xi_mu_centered[5,2,1] 
#>          3.638027e-02         -8.162203e-01          4.028703e-01 
#> xi_mu_centered[6,2,1] xi_mu_centered[7,2,1] xi_mu_centered[8,2,1] 
#>         -5.206452e-01          3.043269e-01         -4.688700e-01 
#> xi_mu_centered[9,2,1] xi_mu_centered[1,3,1] xi_mu_centered[2,3,1] 
#>          3.083806e-04          4.491978e-01          1.811059e-01 
#> xi_mu_centered[3,3,1] xi_mu_centered[4,3,1] xi_mu_centered[5,3,1] 
#>          8.783865e-01         -2.100372e-01          5.542147e-01 
#> xi_mu_centered[6,3,1] xi_mu_centered[7,3,1] xi_mu_centered[8,3,1] 
#>         -7.414588e-01          1.146323e+00          2.957770e-04 
#> xi_mu_centered[9,3,1] xi_mu_centered[1,4,1] xi_mu_centered[2,4,1] 
#>         -1.271150e-04         -1.686505e-01          7.565994e-01 
#> xi_mu_centered[3,4,1] xi_mu_centered[4,4,1] xi_mu_centered[5,4,1] 
#>         -3.353156e-01          3.763328e-01          3.572137e-01 
#> xi_mu_centered[6,4,1] xi_mu_centered[7,4,1] xi_mu_centered[8,4,1] 
#>         -7.051243e-02         -2.761128e-04          1.644785e-04 
#> xi_mu_centered[9,4,1] xi_mu_centered[1,5,1] xi_mu_centered[2,5,1] 
#>         -1.777768e-04         -1.763934e-01          1.159486e+00 
#> xi_mu_centered[3,5,1] xi_mu_centered[4,5,1] xi_mu_centered[5,5,1] 
#>          3.612787e-01          9.530028e-01          2.905843e-01 
#> xi_mu_centered[6,5,1] xi_mu_centered[7,5,1] xi_mu_centered[8,5,1] 
#>         -2.438056e-04          2.361683e-04         -8.053136e-05 
#> xi_mu_centered[9,5,1] xi_mu_centered[1,6,1] xi_mu_centered[2,6,1] 
#>          3.547135e-05          3.074458e-01         -5.738981e-01 
#> xi_mu_centered[3,6,1] xi_mu_centered[4,6,1] xi_mu_centered[5,6,1] 
#>          4.832545e-01          1.276241e-01          5.036634e-05 
#> xi_mu_centered[6,6,1] xi_mu_centered[7,6,1] xi_mu_centered[8,6,1] 
#>         -2.258929e-06         -2.398602e-04          3.300294e-04 
#> xi_mu_centered[9,6,1] xi_mu_centered[1,7,1] xi_mu_centered[2,7,1] 
#>          8.383430e-05          3.768686e-01         -8.238813e-01 
#> xi_mu_centered[3,7,1] xi_mu_centered[4,7,1] xi_mu_centered[5,7,1] 
#>         -2.454151e-01          1.022190e-04         -1.690673e-04 
#> xi_mu_centered[6,7,1] xi_mu_centered[7,7,1] xi_mu_centered[8,7,1] 
#>         -2.439078e-04         -1.466666e-04         -2.338707e-04 
#> xi_mu_centered[9,7,1] xi_mu_centered[1,8,1] xi_mu_centered[2,8,1] 
#>          2.154939e-04         -6.086970e-01         -2.009458e-01 
#> xi_mu_centered[3,8,1] xi_mu_centered[4,8,1] xi_mu_centered[5,8,1] 
#>         -2.582120e-04          2.876734e-04          1.607836e-04 
#> xi_mu_centered[6,8,1] xi_mu_centered[7,8,1] xi_mu_centered[8,8,1] 
#>         -1.791737e-04          5.011532e-05          2.658117e-04 
#> xi_mu_centered[9,8,1] xi_nu_centered[1,1,1] xi_nu_centered[2,1,1] 
#>          2.721904e-04         -3.200012e-04          2.163248e-04 
#> xi_nu_centered[3,1,1] xi_nu_centered[4,1,1] xi_nu_centered[5,1,1] 
#>         -3.141177e-04          1.517164e-04         -2.669093e-04 
#> xi_nu_centered[6,1,1] xi_nu_centered[7,1,1] xi_nu_centered[8,1,1] 
#>          2.525605e-04          2.642398e-04          3.348888e-04 
#> xi_nu_centered[9,1,1] xi_nu_centered[1,2,1] xi_nu_centered[2,2,1] 
#>         -2.102172e-05         -3.384202e-04         -1.853001e-05 
#> xi_nu_centered[3,2,1] xi_nu_centered[4,2,1] xi_nu_centered[5,2,1] 
#>         -2.428940e-04         -2.360179e-04          2.789012e-04 
#> xi_nu_centered[6,2,1] xi_nu_centered[7,2,1] xi_nu_centered[8,2,1] 
#>          1.031099e-04         -1.257747e-04         -1.276166e-04 
#> xi_nu_centered[9,2,1] xi_nu_centered[1,3,1] xi_nu_centered[2,3,1] 
#>          3.280241e-04         -6.207514e-05          2.289363e-04 
#> xi_nu_centered[3,3,1] xi_nu_centered[4,3,1] xi_nu_centered[5,3,1] 
#>          1.631643e-04          1.230416e-04          1.604411e-05 
#> xi_nu_centered[6,3,1] xi_nu_centered[7,3,1] xi_nu_centered[8,3,1] 
#>         -3.493717e-04         -7.928612e-05          1.977904e-04 
#> xi_nu_centered[9,3,1] xi_nu_centered[1,4,1] xi_nu_centered[2,4,1] 
#>          2.127277e-04          2.856517e-04          1.345493e-05 
#> xi_nu_centered[3,4,1] xi_nu_centered[4,4,1] xi_nu_centered[5,4,1] 
#>         -1.245130e-04         -2.665561e-04         -3.319035e-04 
#> xi_nu_centered[6,4,1] xi_nu_centered[7,4,1] xi_nu_centered[8,4,1] 
#>         -3.242039e-04         -1.879090e-04         -3.471291e-05 
#> xi_nu_centered[9,4,1] xi_nu_centered[1,5,1] xi_nu_centered[2,5,1] 
#>         -1.092930e-04          2.660951e-04         -4.568021e-05 
#> xi_nu_centered[3,5,1] xi_nu_centered[4,5,1] xi_nu_centered[5,5,1] 
#>          1.348875e-04          3.058873e-04         -1.631725e-04 
#> xi_nu_centered[6,5,1] xi_nu_centered[7,5,1] xi_nu_centered[8,5,1] 
#>          1.764157e-04         -1.215114e-04          3.602981e-04 
#> xi_nu_centered[9,5,1] xi_nu_centered[1,6,1] xi_nu_centered[2,6,1] 
#>         -2.932238e-04          1.816927e-04         -1.783128e-04 
#> xi_nu_centered[3,6,1] xi_nu_centered[4,6,1] xi_nu_centered[5,6,1] 
#>         -1.154444e-04          1.081631e-04         -3.159012e-04 
#> xi_nu_centered[6,6,1] xi_nu_centered[7,6,1] xi_nu_centered[8,6,1] 
#>          2.909444e-04          2.817376e-04          2.745257e-04 
#> xi_nu_centered[9,6,1] xi_nu_centered[1,7,1] xi_nu_centered[2,7,1] 
#>         -3.244793e-04          2.112667e-04          7.209258e-05 
#> xi_nu_centered[3,7,1] xi_nu_centered[4,7,1] xi_nu_centered[5,7,1] 
#>          3.069485e-04          2.865064e-04          7.082495e-05 
#> xi_nu_centered[6,7,1] xi_nu_centered[7,7,1] xi_nu_centered[8,7,1] 
#>         -1.709953e-04         -4.286196e-05         -2.087583e-04 
#> xi_nu_centered[9,7,1] xi_nu_centered[1,8,1] xi_nu_centered[2,8,1] 
#>         -8.377315e-05         -4.066652e-05         -1.234871e-04 
#> xi_nu_centered[3,8,1] xi_nu_centered[4,8,1] xi_nu_centered[5,8,1] 
#>          5.316986e-05         -1.809430e-04         -6.542553e-07 
#> xi_nu_centered[6,8,1] xi_nu_centered[7,8,1] xi_nu_centered[8,8,1] 
#>         -1.888028e-04          1.875214e-04          2.073268e-05 
#> xi_nu_centered[9,8,1]                  r[1]             mu_0[1,1] 
#>         -1.962782e-04          3.247471e+00         -1.148584e-01 
#>             mu_0[2,1]             mu_0[3,1]             mu_0[4,1] 
#>         -4.976166e-01         -4.077122e-01         -5.009056e-01 
#>             mu_0[5,1]             mu_0[6,1]             mu_0[7,1] 
#>         -7.708262e-02         -1.525706e-02         -6.541346e-03 
#>             mu_0[8,1]             nu_0[1,1]             nu_0[2,1] 
#>          2.118975e-01         -1.608361e-01         -1.726782e-01 
#>             nu_0[3,1]             nu_0[4,1]             nu_0[5,1] 
#>         -1.698956e-01         -1.727601e-01         -1.596541e-01 
#>             nu_0[6,1]             nu_0[7,1]             nu_0[8,1] 
#>         -1.578086e-01         -1.575505e-01         -1.508189e-01 
#>          xi_mu[1,1,1]          xi_mu[2,1,1]          xi_mu[3,1,1] 
#>         -1.665940e+00          1.403493e-01          2.204959e+00 
#>          xi_mu[4,1,1]          xi_mu[5,1,1]          xi_mu[6,1,1] 
#>         -2.311494e+00         -4.481980e-01          7.659246e-01 
#>          xi_mu[7,1,1]          xi_mu[8,1,1]          xi_mu[9,1,1] 
#>          3.284845e+00          5.607350e-01          2.361805e+00 
#>          xi_mu[1,2,1]          xi_mu[2,2,1]          xi_mu[3,2,1] 
#>         -5.388320e-01          3.056904e+00          1.319162e-01 
#>          xi_mu[4,2,1]          xi_mu[5,2,1]          xi_mu[6,2,1] 
#>         -2.959644e+00          1.460822e+00         -1.887878e+00 
#>          xi_mu[7,2,1]          xi_mu[8,2,1]          xi_mu[9,2,1] 
#>          1.103500e+00         -1.700139e+00          1.118199e-03 
#>          xi_mu[1,3,1]          xi_mu[2,3,1]          xi_mu[3,3,1] 
#>          1.628807e+00          6.566966e-01          3.185061e+00 
#>          xi_mu[4,3,1]          xi_mu[5,3,1]          xi_mu[6,3,1] 
#>         -7.616023e-01          2.009602e+00         -2.688556e+00 
#>          xi_mu[7,3,1]          xi_mu[8,3,1]          xi_mu[9,3,1] 
#>          4.156610e+00          1.072498e-03         -4.609234e-04 
#>          xi_mu[1,4,1]          xi_mu[2,4,1]          xi_mu[3,4,1] 
#>         -6.115328e-01          2.743456e+00         -1.215866e+00 
#>          xi_mu[4,4,1]          xi_mu[5,4,1]          xi_mu[6,4,1] 
#>          1.364596e+00          1.295270e+00         -2.556806e-01 
#>          xi_mu[7,4,1]          xi_mu[8,4,1]          xi_mu[9,4,1] 
#>         -1.001195e-03          5.964048e-04         -6.446252e-04 
#>          xi_mu[1,5,1]          xi_mu[2,5,1]          xi_mu[3,5,1] 
#>         -6.396089e-01          4.204338e+00          1.310009e+00 
#>          xi_mu[4,5,1]          xi_mu[5,5,1]          xi_mu[6,5,1] 
#>          3.455622e+00          1.053669e+00         -8.840478e-04 
#>          xi_mu[7,5,1]          xi_mu[8,5,1]          xi_mu[9,5,1] 
#>          8.563548e-04         -2.920096e-04          1.286204e-04 
#>          xi_mu[1,6,1]          xi_mu[2,6,1]          xi_mu[3,6,1] 
#>          1.114809e+00         -2.080975e+00          1.752298e+00 
#>          xi_mu[4,6,1]          xi_mu[5,6,1]          xi_mu[6,6,1] 
#>          4.627694e-01          1.826302e-04         -8.190956e-06 
#>          xi_mu[7,6,1]          xi_mu[8,6,1]          xi_mu[9,6,1] 
#>         -8.697417e-04          1.196698e-03          3.039862e-04 
#>          xi_mu[1,7,1]          xi_mu[2,7,1]          xi_mu[3,7,1] 
#>          1.366539e+00         -2.987423e+00         -8.898838e-01 
#>          xi_mu[4,7,1]          xi_mu[5,7,1]          xi_mu[6,7,1] 
#>          3.706496e-04         -6.130439e-04         -8.844184e-04 
#>          xi_mu[7,7,1]          xi_mu[8,7,1]          xi_mu[9,7,1] 
#>         -5.318184e-04         -8.480234e-04          7.813886e-04 
#>          xi_mu[1,8,1]          xi_mu[2,8,1]          xi_mu[3,8,1] 
#>         -2.207157e+00         -7.286368e-01         -9.362859e-04 
#>          xi_mu[4,8,1]          xi_mu[5,8,1]          xi_mu[6,8,1] 
#>          1.043114e-03          5.830071e-04         -6.496901e-04 
#>          xi_mu[7,8,1]          xi_mu[8,8,1]          xi_mu[9,8,1] 
#>          1.817199e-04          9.638429e-04          9.869720e-04 
#>          xi_nu[1,1,1]          xi_nu[2,1,1]          xi_nu[3,1,1] 
#>         -1.374210e-05          9.289830e-06         -1.348944e-05 
#>          xi_nu[4,1,1]          xi_nu[5,1,1]          xi_nu[6,1,1] 
#>          6.515291e-06         -1.146212e-05          1.084593e-05 
#>          xi_nu[7,1,1]          xi_nu[8,1,1]          xi_nu[9,1,1] 
#>          1.134749e-05          1.438143e-05         -9.027544e-07 
#>          xi_nu[1,2,1]          xi_nu[2,2,1]          xi_nu[3,2,1] 
#>         -1.453308e-05         -7.957507e-07         -1.043082e-05 
#>          xi_nu[4,2,1]          xi_nu[5,2,1]          xi_nu[6,2,1] 
#>         -1.013553e-05          1.197710e-05          4.427941e-06 
#>          xi_nu[7,2,1]          xi_nu[8,2,1]          xi_nu[9,2,1] 
#>         -5.401255e-06         -5.480353e-06          1.408663e-05 
#>          xi_nu[1,3,1]          xi_nu[2,3,1]          xi_nu[3,3,1] 
#>         -2.665748e-06          9.831417e-06          7.006909e-06 
#>          xi_nu[4,3,1]          xi_nu[5,3,1]          xi_nu[6,3,1] 
#>          5.283883e-06          6.889966e-07         -1.500338e-05 
#>          xi_nu[7,3,1]          xi_nu[8,3,1]          xi_nu[9,3,1] 
#>         -3.404855e-06          8.493888e-06          9.135354e-06 
#>          xi_nu[1,4,1]          xi_nu[2,4,1]          xi_nu[3,4,1] 
#>          1.226700e-05          5.778070e-07         -5.347073e-06 
#>          xi_nu[4,4,1]          xi_nu[5,4,1]          xi_nu[6,4,1] 
#>         -1.144696e-05         -1.425323e-05         -1.392258e-05 
#>          xi_nu[7,4,1]          xi_nu[8,4,1]          xi_nu[9,4,1] 
#>         -8.069543e-06         -1.490708e-06         -4.693467e-06 
#>          xi_nu[1,5,1]          xi_nu[2,5,1]          xi_nu[3,5,1] 
#>          1.142716e-05         -1.961686e-06          5.792596e-06 
#>          xi_nu[4,5,1]          xi_nu[5,5,1]          xi_nu[6,5,1] 
#>          1.313599e-05         -7.007263e-06          7.575976e-06 
#>          xi_nu[7,5,1]          xi_nu[8,5,1]          xi_nu[9,5,1] 
#>         -5.218173e-06          1.547260e-05         -1.259217e-05 
#>          xi_nu[1,6,1]          xi_nu[2,6,1]          xi_nu[3,6,1] 
#>          7.802592e-06         -7.657445e-06         -4.957633e-06 
#>          xi_nu[4,6,1]          xi_nu[5,6,1]          xi_nu[6,6,1] 
#>          4.644943e-06         -1.356603e-05          1.249428e-05 
#>          xi_nu[7,6,1]          xi_nu[8,6,1]          xi_nu[9,6,1] 
#>          1.209891e-05          1.178920e-05         -1.393440e-05 
#>          xi_nu[1,7,1]          xi_nu[2,7,1]          xi_nu[3,7,1] 
#>          9.072613e-06          3.095936e-06          1.318157e-05 
#>          xi_nu[4,7,1]          xi_nu[5,7,1]          xi_nu[6,7,1] 
#>          1.230370e-05          3.041499e-06         -7.343204e-06 
#>          xi_nu[7,7,1]          xi_nu[8,7,1]          xi_nu[9,7,1] 
#>         -1.840659e-06         -8.964895e-06         -3.597545e-06 
#>          xi_nu[1,8,1]          xi_nu[2,8,1]          xi_nu[3,8,1] 
#>         -1.746379e-06         -5.303016e-06          2.283321e-06 
#>          xi_nu[4,8,1]          xi_nu[5,8,1]          xi_nu[6,8,1] 
#>         -7.770395e-06         -2.809627e-08         -8.107928e-06 
#>          xi_nu[7,8,1]          xi_nu[8,8,1]          xi_nu[9,8,1] 
#>          8.052899e-06          8.903420e-07         -8.428948e-06 
#>           lambda[1,1]           lambda[2,1]           lambda[3,1] 
#>         -2.756945e-01         -6.702948e-01         -5.776078e-01 
#>           lambda[4,1]           lambda[5,1]           lambda[6,1] 
#>         -6.736658e-01         -2.367367e-01         -1.730657e-01 
#>           lambda[7,1]           lambda[8,1]           lambda[1,2] 
#>         -1.640918e-01          6.107859e-02         -1.941634e+00 
#>           lambda[2,2]           lambda[3,2]           lambda[4,2] 
#>         -1.209127e+00          1.051200e+00         -1.285199e+00 
#>           lambda[5,2]           lambda[6,2]           lambda[7,2] 
#>         -8.763456e-01          9.417437e-01          1.202447e+00 
#>           lambda[8,2]           lambda[1,3]           lambda[2,3] 
#>         -2.146079e+00         -1.801285e+00          1.847777e+00 
#>           lambda[3,3]           lambda[4,3]           lambda[5,3] 
#>          1.707896e+00          1.458258e+00          3.327993e+00 
#>           lambda[6,3]           lambda[7,3]           lambda[8,3] 
#>         -1.139231e+00         -1.784976e+00         -2.874715e+00 
#>           lambda[1,4]           lambda[2,4]           lambda[3,4] 
#>          4.036746e-01          1.979693e+00          4.892957e+00 
#>           lambda[4,4]           lambda[5,4]           lambda[6,4] 
#>          2.423914e-01          4.638002e+00          6.130669e-01 
#>           lambda[7,4]           lambda[8,4]           lambda[1,5] 
#>         -2.674860e+00         -2.875652e+00         -1.907820e+00 
#>           lambda[2,5]           lambda[3,5]           lambda[4,5] 
#>         -9.799509e-01          4.131355e+00          1.606988e+00 
#>           lambda[5,5]           lambda[6,5]           lambda[7,5] 
#>          8.093624e+00          1.075836e+00         -2.674489e+00 
#>           lambda[8,5]           lambda[1,6]           lambda[2,6] 
#>         -2.874608e+00         -2.356018e+00          4.808712e-01 
#>           lambda[3,6]           lambda[4,6]           lambda[5,6] 
#>          6.140957e+00          2.902257e+00          9.147293e+00 
#>           lambda[6,6]           lambda[7,6]           lambda[8,6] 
#>          1.076019e+00         -2.675102e+00         -2.874025e+00 
#>           lambda[1,7]           lambda[2,7]           lambda[3,7] 
#>         -1.590093e+00         -1.407007e+00          3.452401e+00 
#>           lambda[4,7]           lambda[5,7]           lambda[6,7] 
#>          2.646577e+00          9.146409e+00          1.076011e+00 
#>           lambda[7,7]           lambda[8,7]           lambda[1,8] 
#>         -2.675986e+00         -2.874675e+00          1.694752e+00 
#>           lambda[2,8]           lambda[3,8]           lambda[4,8] 
#>         -3.035066e-01          7.609011e+00          2.645576e+00 
#>           lambda[5,8]           lambda[6,8]           lambda[7,8] 
#>          9.147265e+00          1.075141e+00         -2.676518e+00 
#>           lambda[8,8]           lambda[1,9]           lambda[2,9] 
#>         -2.874493e+00          2.255487e+00         -2.003646e+00 
#>           lambda[3,9]           lambda[4,9]           lambda[5,9] 
#>          7.610083e+00          2.646172e+00          9.146973e+00 
#>           lambda[6,9]           lambda[7,9]           lambda[8,9] 
#>          1.076338e+00         -2.677366e+00         -2.873530e+00 
#>          lambda[1,10]          lambda[2,10]          lambda[3,10] 
#>          4.617292e+00         -2.002528e+00          7.609622e+00 
#>          lambda[4,10]          lambda[5,10]          lambda[6,10] 
#>          2.645527e+00          9.147102e+00          1.076642e+00 
#>          lambda[7,10]          lambda[8,10]        lambda_mean[1] 
#>         -2.676585e+00         -2.872543e+00         -2.756945e-01 
#>        lambda_mean[2]        lambda_mean[3]        lambda_mean[4] 
#>         -6.702948e-01         -5.776078e-01         -6.736658e-01 
#>        lambda_mean[5]        lambda_mean[6]        lambda_mean[7] 
#>         -2.367367e-01         -1.730657e-01         -1.640918e-01 
#>        lambda_mean[8]        lambda_mean[9]       lambda_mean[10] 
#>          6.107859e-02         -1.941634e+00         -1.209127e+00 
#>       lambda_mean[11]       lambda_mean[12]       lambda_mean[13] 
#>          1.051200e+00         -1.285199e+00         -8.763456e-01 
#>       lambda_mean[14]       lambda_mean[15]       lambda_mean[16] 
#>          9.417437e-01          1.202447e+00         -2.146079e+00 
#>       lambda_mean[17]       lambda_mean[18]       lambda_mean[19] 
#>         -1.801285e+00          1.847777e+00          1.707896e+00 
#>       lambda_mean[20]       lambda_mean[21]       lambda_mean[22] 
#>          1.458258e+00          3.327993e+00         -1.139231e+00 
#>       lambda_mean[23]       lambda_mean[24]       lambda_mean[25] 
#>         -1.784976e+00         -2.874715e+00          4.036746e-01 
#>       lambda_mean[26]       lambda_mean[27]       lambda_mean[28] 
#>          1.979693e+00          4.892957e+00          2.423914e-01 
#>       lambda_mean[29]       lambda_mean[30]       lambda_mean[31] 
#>          4.638002e+00          6.130669e-01         -2.674860e+00 
#>       lambda_mean[32]       lambda_mean[33]       lambda_mean[34] 
#>         -1.907820e+00         -9.799509e-01          4.131355e+00 
#>       lambda_mean[35]       lambda_mean[36]       lambda_mean[37] 
#>          1.606988e+00          8.093624e+00          1.075836e+00 
#>       lambda_mean[38]       lambda_mean[39]       lambda_mean[40] 
#>         -2.356018e+00          4.808712e-01          6.140957e+00 
#>       lambda_mean[41]       lambda_mean[42]       lambda_mean[43] 
#>          2.902257e+00          9.147293e+00         -1.590093e+00 
#>       lambda_mean[44]       lambda_mean[45]       lambda_mean[46] 
#>         -1.407007e+00          3.452401e+00          2.646577e+00 
#>       lambda_mean[47]       lambda_mean[48]       lambda_mean[49] 
#>          1.694752e+00         -3.035066e-01          7.609011e+00 
#>       lambda_mean[50]       lambda_mean[51]       lambda_mean[52] 
#>          2.255487e+00         -2.003646e+00          4.617292e+00 
#>                lprior 
#>         -1.726217e+02 
#> 
#> $model$value
#> [1] -281.8103
#> 
#> $model$return_code
#> [1] 0
#> 
#> $model$theta_tilde
#>      mu_0_centered[1,1] mu_0_centered[2,1] mu_0_centered[3,1]
#> [1,]         0.05653347         -0.4022521         -0.2944901
#>      mu_0_centered[4,1] mu_0_centered[5,1] mu_0_centered[6,1]
#> [1,]         -0.4061945          0.1018126          0.1759186
#>      mu_0_centered[7,1] mu_0_centered[8,1] nu_0_centered[1,1]
#> [1,]          0.1863656           0.448193         0.01028474
#>      nu_0_centered[2,1] nu_0_centered[3,1] nu_0_centered[4,1]
#> [1,]        -0.07051117        -0.05152598        -0.07107033
#>      nu_0_centered[5,1] nu_0_centered[6,1] nu_0_centered[7,1]
#> [1,]         0.01834992         0.03094108         0.03270211
#>      nu_0_centered[8,1]  mu_0_mean  nu_0_mean   mu_0_sd   nu_0_sd    mu_sd
#> [1,]         0.07863037 -0.1620234 -0.1623436 0.8342856 0.1465672 3.626036
#>           nu_sd xi_mu_centered[1,1,1] xi_mu_centered[2,1,1]
#> [1,] 0.04294389            -0.4594383            0.03870599
#>      xi_mu_centered[3,1,1] xi_mu_centered[4,1,1] xi_mu_centered[5,1,1]
#> [1,]             0.6080909            -0.6374715            -0.1236055
#>      xi_mu_centered[6,1,1] xi_mu_centered[7,1,1] xi_mu_centered[8,1,1]
#> [1,]             0.2112292             0.9059053             0.1546413
#>      xi_mu_centered[9,1,1] xi_mu_centered[1,2,1] xi_mu_centered[2,2,1]
#> [1,]             0.6513463            -0.1486009             0.8430429
#>      xi_mu_centered[3,2,1] xi_mu_centered[4,2,1] xi_mu_centered[5,2,1]
#> [1,]            0.03638027            -0.8162203             0.4028703
#>      xi_mu_centered[6,2,1] xi_mu_centered[7,2,1] xi_mu_centered[8,2,1]
#> [1,]            -0.5206452             0.3043269              -0.46887
#>      xi_mu_centered[9,2,1] xi_mu_centered[1,3,1] xi_mu_centered[2,3,1]
#> [1,]          0.0003083806             0.4491978             0.1811059
#>      xi_mu_centered[3,3,1] xi_mu_centered[4,3,1] xi_mu_centered[5,3,1]
#> [1,]             0.8783865            -0.2100372             0.5542147
#>      xi_mu_centered[6,3,1] xi_mu_centered[7,3,1] xi_mu_centered[8,3,1]
#> [1,]            -0.7414588              1.146323           0.000295777
#>      xi_mu_centered[9,3,1] xi_mu_centered[1,4,1] xi_mu_centered[2,4,1]
#> [1,]          -0.000127115            -0.1686505             0.7565994
#>      xi_mu_centered[3,4,1] xi_mu_centered[4,4,1] xi_mu_centered[5,4,1]
#> [1,]            -0.3353156             0.3763328             0.3572137
#>      xi_mu_centered[6,4,1] xi_mu_centered[7,4,1] xi_mu_centered[8,4,1]
#> [1,]           -0.07051243         -0.0002761128          0.0001644785
#>      xi_mu_centered[9,4,1] xi_mu_centered[1,5,1] xi_mu_centered[2,5,1]
#> [1,]         -0.0001777768            -0.1763934              1.159486
#>      xi_mu_centered[3,5,1] xi_mu_centered[4,5,1] xi_mu_centered[5,5,1]
#> [1,]             0.3612787             0.9530028             0.2905843
#>      xi_mu_centered[6,5,1] xi_mu_centered[7,5,1] xi_mu_centered[8,5,1]
#> [1,]         -0.0002438056          0.0002361683         -8.053136e-05
#>      xi_mu_centered[9,5,1] xi_mu_centered[1,6,1] xi_mu_centered[2,6,1]
#> [1,]          3.547135e-05             0.3074458            -0.5738981
#>      xi_mu_centered[3,6,1] xi_mu_centered[4,6,1] xi_mu_centered[5,6,1]
#> [1,]             0.4832545             0.1276241          5.036634e-05
#>      xi_mu_centered[6,6,1] xi_mu_centered[7,6,1] xi_mu_centered[8,6,1]
#> [1,]         -2.258929e-06         -0.0002398602          0.0003300294
#>      xi_mu_centered[9,6,1] xi_mu_centered[1,7,1] xi_mu_centered[2,7,1]
#> [1,]           8.38343e-05             0.3768686            -0.8238813
#>      xi_mu_centered[3,7,1] xi_mu_centered[4,7,1] xi_mu_centered[5,7,1]
#> [1,]            -0.2454151           0.000102219         -0.0001690673
#>      xi_mu_centered[6,7,1] xi_mu_centered[7,7,1] xi_mu_centered[8,7,1]
#> [1,]         -0.0002439078         -0.0001466666         -0.0002338707
#>      xi_mu_centered[9,7,1] xi_mu_centered[1,8,1] xi_mu_centered[2,8,1]
#> [1,]          0.0002154939             -0.608697            -0.2009458
#>      xi_mu_centered[3,8,1] xi_mu_centered[4,8,1] xi_mu_centered[5,8,1]
#> [1,]          -0.000258212          0.0002876734          0.0001607836
#>      xi_mu_centered[6,8,1] xi_mu_centered[7,8,1] xi_mu_centered[8,8,1]
#> [1,]         -0.0001791737          5.011532e-05          0.0002658117
#>      xi_mu_centered[9,8,1] xi_nu_centered[1,1,1] xi_nu_centered[2,1,1]
#> [1,]          0.0002721904         -0.0003200012          0.0002163248
#>      xi_nu_centered[3,1,1] xi_nu_centered[4,1,1] xi_nu_centered[5,1,1]
#> [1,]         -0.0003141177          0.0001517164         -0.0002669093
#>      xi_nu_centered[6,1,1] xi_nu_centered[7,1,1] xi_nu_centered[8,1,1]
#> [1,]          0.0002525605          0.0002642398          0.0003348888
#>      xi_nu_centered[9,1,1] xi_nu_centered[1,2,1] xi_nu_centered[2,2,1]
#> [1,]         -2.102172e-05         -0.0003384202         -1.853001e-05
#>      xi_nu_centered[3,2,1] xi_nu_centered[4,2,1] xi_nu_centered[5,2,1]
#> [1,]          -0.000242894         -0.0002360179          0.0002789012
#>      xi_nu_centered[6,2,1] xi_nu_centered[7,2,1] xi_nu_centered[8,2,1]
#> [1,]          0.0001031099         -0.0001257747         -0.0001276166
#>      xi_nu_centered[9,2,1] xi_nu_centered[1,3,1] xi_nu_centered[2,3,1]
#> [1,]          0.0003280241         -6.207514e-05          0.0002289363
#>      xi_nu_centered[3,3,1] xi_nu_centered[4,3,1] xi_nu_centered[5,3,1]
#> [1,]          0.0001631643          0.0001230416          1.604411e-05
#>      xi_nu_centered[6,3,1] xi_nu_centered[7,3,1] xi_nu_centered[8,3,1]
#> [1,]         -0.0003493717         -7.928612e-05          0.0001977904
#>      xi_nu_centered[9,3,1] xi_nu_centered[1,4,1] xi_nu_centered[2,4,1]
#> [1,]          0.0002127277          0.0002856517          1.345493e-05
#>      xi_nu_centered[3,4,1] xi_nu_centered[4,4,1] xi_nu_centered[5,4,1]
#> [1,]          -0.000124513         -0.0002665561         -0.0003319035
#>      xi_nu_centered[6,4,1] xi_nu_centered[7,4,1] xi_nu_centered[8,4,1]
#> [1,]         -0.0003242039          -0.000187909         -3.471291e-05
#>      xi_nu_centered[9,4,1] xi_nu_centered[1,5,1] xi_nu_centered[2,5,1]
#> [1,]          -0.000109293          0.0002660951         -4.568021e-05
#>      xi_nu_centered[3,5,1] xi_nu_centered[4,5,1] xi_nu_centered[5,5,1]
#> [1,]          0.0001348875          0.0003058873         -0.0001631725
#>      xi_nu_centered[6,5,1] xi_nu_centered[7,5,1] xi_nu_centered[8,5,1]
#> [1,]          0.0001764157         -0.0001215114          0.0003602981
#>      xi_nu_centered[9,5,1] xi_nu_centered[1,6,1] xi_nu_centered[2,6,1]
#> [1,]         -0.0002932238          0.0001816927         -0.0001783128
#>      xi_nu_centered[3,6,1] xi_nu_centered[4,6,1] xi_nu_centered[5,6,1]
#> [1,]         -0.0001154444          0.0001081631         -0.0003159012
#>      xi_nu_centered[6,6,1] xi_nu_centered[7,6,1] xi_nu_centered[8,6,1]
#> [1,]          0.0002909444          0.0002817376          0.0002745257
#>      xi_nu_centered[9,6,1] xi_nu_centered[1,7,1] xi_nu_centered[2,7,1]
#> [1,]         -0.0003244793          0.0002112667          7.209258e-05
#>      xi_nu_centered[3,7,1] xi_nu_centered[4,7,1] xi_nu_centered[5,7,1]
#> [1,]          0.0003069485          0.0002865064          7.082495e-05
#>      xi_nu_centered[6,7,1] xi_nu_centered[7,7,1] xi_nu_centered[8,7,1]
#> [1,]         -0.0001709953         -4.286196e-05         -0.0002087583
#>      xi_nu_centered[9,7,1] xi_nu_centered[1,8,1] xi_nu_centered[2,8,1]
#> [1,]         -8.377315e-05         -4.066652e-05         -0.0001234871
#>      xi_nu_centered[3,8,1] xi_nu_centered[4,8,1] xi_nu_centered[5,8,1]
#> [1,]          5.316986e-05          -0.000180943         -6.542553e-07
#>      xi_nu_centered[6,8,1] xi_nu_centered[7,8,1] xi_nu_centered[8,8,1]
#> [1,]         -0.0001888028          0.0001875214          2.073268e-05
#>      xi_nu_centered[9,8,1]     r[1]  mu_0[1,1]  mu_0[2,1]  mu_0[3,1]  mu_0[4,1]
#> [1,]         -0.0001962782 3.247471 -0.1148584 -0.4976166 -0.4077122 -0.5009056
#>        mu_0[5,1]   mu_0[6,1]    mu_0[7,1] mu_0[8,1]  nu_0[1,1]  nu_0[2,1]
#> [1,] -0.07708262 -0.01525706 -0.006541346 0.2118975 -0.1608361 -0.1726782
#>       nu_0[3,1]  nu_0[4,1]  nu_0[5,1]  nu_0[6,1]  nu_0[7,1]  nu_0[8,1]
#> [1,] -0.1698956 -0.1727601 -0.1596541 -0.1578086 -0.1575505 -0.1508189
#>      xi_mu[1,1,1] xi_mu[2,1,1] xi_mu[3,1,1] xi_mu[4,1,1] xi_mu[5,1,1]
#> [1,]     -1.66594    0.1403493     2.204959    -2.311494    -0.448198
#>      xi_mu[6,1,1] xi_mu[7,1,1] xi_mu[8,1,1] xi_mu[9,1,1] xi_mu[1,2,1]
#> [1,]    0.7659246     3.284845     0.560735     2.361805    -0.538832
#>      xi_mu[2,2,1] xi_mu[3,2,1] xi_mu[4,2,1] xi_mu[5,2,1] xi_mu[6,2,1]
#> [1,]     3.056904    0.1319162    -2.959644     1.460822    -1.887878
#>      xi_mu[7,2,1] xi_mu[8,2,1] xi_mu[9,2,1] xi_mu[1,3,1] xi_mu[2,3,1]
#> [1,]       1.1035    -1.700139  0.001118199     1.628807    0.6566966
#>      xi_mu[3,3,1] xi_mu[4,3,1] xi_mu[5,3,1] xi_mu[6,3,1] xi_mu[7,3,1]
#> [1,]     3.185061   -0.7616023     2.009602    -2.688556      4.15661
#>      xi_mu[8,3,1]  xi_mu[9,3,1] xi_mu[1,4,1] xi_mu[2,4,1] xi_mu[3,4,1]
#> [1,]  0.001072498 -0.0004609234   -0.6115328     2.743456    -1.215866
#>      xi_mu[4,4,1] xi_mu[5,4,1] xi_mu[6,4,1] xi_mu[7,4,1] xi_mu[8,4,1]
#> [1,]     1.364596      1.29527   -0.2556806 -0.001001195 0.0005964048
#>       xi_mu[9,4,1] xi_mu[1,5,1] xi_mu[2,5,1] xi_mu[3,5,1] xi_mu[4,5,1]
#> [1,] -0.0006446252   -0.6396089     4.204338     1.310009     3.455622
#>      xi_mu[5,5,1]  xi_mu[6,5,1] xi_mu[7,5,1]  xi_mu[8,5,1] xi_mu[9,5,1]
#> [1,]     1.053669 -0.0008840478 0.0008563548 -0.0002920096 0.0001286204
#>      xi_mu[1,6,1] xi_mu[2,6,1] xi_mu[3,6,1] xi_mu[4,6,1] xi_mu[5,6,1]
#> [1,]     1.114809    -2.080975     1.752298    0.4627694 0.0001826302
#>       xi_mu[6,6,1]  xi_mu[7,6,1] xi_mu[8,6,1] xi_mu[9,6,1] xi_mu[1,7,1]
#> [1,] -8.190956e-06 -0.0008697417  0.001196698 0.0003039862     1.366539
#>      xi_mu[2,7,1] xi_mu[3,7,1] xi_mu[4,7,1]  xi_mu[5,7,1]  xi_mu[6,7,1]
#> [1,]    -2.987423   -0.8898838 0.0003706496 -0.0006130439 -0.0008844184
#>       xi_mu[7,7,1]  xi_mu[8,7,1] xi_mu[9,7,1] xi_mu[1,8,1] xi_mu[2,8,1]
#> [1,] -0.0005318184 -0.0008480234 0.0007813886    -2.207157   -0.7286368
#>       xi_mu[3,8,1] xi_mu[4,8,1] xi_mu[5,8,1]  xi_mu[6,8,1] xi_mu[7,8,1]
#> [1,] -0.0009362859  0.001043114 0.0005830071 -0.0006496901 0.0001817199
#>      xi_mu[8,8,1] xi_mu[9,8,1] xi_nu[1,1,1] xi_nu[2,1,1]  xi_nu[3,1,1]
#> [1,] 0.0009638429  0.000986972 -1.37421e-05  9.28983e-06 -1.348944e-05
#>      xi_nu[4,1,1]  xi_nu[5,1,1] xi_nu[6,1,1] xi_nu[7,1,1] xi_nu[8,1,1]
#> [1,] 6.515291e-06 -1.146212e-05 1.084593e-05 1.134749e-05 1.438143e-05
#>       xi_nu[9,1,1]  xi_nu[1,2,1]  xi_nu[2,2,1]  xi_nu[3,2,1]  xi_nu[4,2,1]
#> [1,] -9.027544e-07 -1.453308e-05 -7.957507e-07 -1.043082e-05 -1.013553e-05
#>      xi_nu[5,2,1] xi_nu[6,2,1]  xi_nu[7,2,1]  xi_nu[8,2,1] xi_nu[9,2,1]
#> [1,]  1.19771e-05 4.427941e-06 -5.401255e-06 -5.480353e-06 1.408663e-05
#>       xi_nu[1,3,1] xi_nu[2,3,1] xi_nu[3,3,1] xi_nu[4,3,1] xi_nu[5,3,1]
#> [1,] -2.665748e-06 9.831417e-06 7.006909e-06 5.283883e-06 6.889966e-07
#>       xi_nu[6,3,1]  xi_nu[7,3,1] xi_nu[8,3,1] xi_nu[9,3,1] xi_nu[1,4,1]
#> [1,] -1.500338e-05 -3.404855e-06 8.493888e-06 9.135354e-06   1.2267e-05
#>      xi_nu[2,4,1]  xi_nu[3,4,1]  xi_nu[4,4,1]  xi_nu[5,4,1]  xi_nu[6,4,1]
#> [1,]  5.77807e-07 -5.347073e-06 -1.144696e-05 -1.425323e-05 -1.392258e-05
#>       xi_nu[7,4,1]  xi_nu[8,4,1]  xi_nu[9,4,1] xi_nu[1,5,1]  xi_nu[2,5,1]
#> [1,] -8.069543e-06 -1.490708e-06 -4.693467e-06 1.142716e-05 -1.961686e-06
#>      xi_nu[3,5,1] xi_nu[4,5,1]  xi_nu[5,5,1] xi_nu[6,5,1]  xi_nu[7,5,1]
#> [1,] 5.792596e-06 1.313599e-05 -7.007263e-06 7.575976e-06 -5.218173e-06
#>      xi_nu[8,5,1]  xi_nu[9,5,1] xi_nu[1,6,1]  xi_nu[2,6,1]  xi_nu[3,6,1]
#> [1,]  1.54726e-05 -1.259217e-05 7.802592e-06 -7.657445e-06 -4.957633e-06
#>      xi_nu[4,6,1]  xi_nu[5,6,1] xi_nu[6,6,1] xi_nu[7,6,1] xi_nu[8,6,1]
#> [1,] 4.644943e-06 -1.356603e-05 1.249428e-05 1.209891e-05  1.17892e-05
#>      xi_nu[9,6,1] xi_nu[1,7,1] xi_nu[2,7,1] xi_nu[3,7,1] xi_nu[4,7,1]
#> [1,] -1.39344e-05 9.072613e-06 3.095936e-06 1.318157e-05  1.23037e-05
#>      xi_nu[5,7,1]  xi_nu[6,7,1]  xi_nu[7,7,1]  xi_nu[8,7,1]  xi_nu[9,7,1]
#> [1,] 3.041499e-06 -7.343204e-06 -1.840659e-06 -8.964895e-06 -3.597545e-06
#>       xi_nu[1,8,1]  xi_nu[2,8,1] xi_nu[3,8,1]  xi_nu[4,8,1]  xi_nu[5,8,1]
#> [1,] -1.746379e-06 -5.303016e-06 2.283321e-06 -7.770395e-06 -2.809627e-08
#>       xi_nu[6,8,1] xi_nu[7,8,1] xi_nu[8,8,1]  xi_nu[9,8,1] lambda[1,1]
#> [1,] -8.107928e-06 8.052899e-06  8.90342e-07 -8.428948e-06  -0.2756945
#>      lambda[2,1] lambda[3,1] lambda[4,1] lambda[5,1] lambda[6,1] lambda[7,1]
#> [1,]  -0.6702948  -0.5776078  -0.6736658  -0.2367367  -0.1730657  -0.1640918
#>      lambda[8,1] lambda[1,2] lambda[2,2] lambda[3,2] lambda[4,2] lambda[5,2]
#> [1,]  0.06107859   -1.941634   -1.209127      1.0512   -1.285199  -0.8763456
#>      lambda[6,2] lambda[7,2] lambda[8,2] lambda[1,3] lambda[2,3] lambda[3,3]
#> [1,]   0.9417437    1.202447   -2.146079   -1.801285    1.847777    1.707896
#>      lambda[4,3] lambda[5,3] lambda[6,3] lambda[7,3] lambda[8,3] lambda[1,4]
#> [1,]    1.458258    3.327993   -1.139231   -1.784976   -2.874715   0.4036746
#>      lambda[2,4] lambda[3,4] lambda[4,4] lambda[5,4] lambda[6,4] lambda[7,4]
#> [1,]    1.979693    4.892957   0.2423914    4.638002   0.6130669    -2.67486
#>      lambda[8,4] lambda[1,5] lambda[2,5] lambda[3,5] lambda[4,5] lambda[5,5]
#> [1,]   -2.875652    -1.90782  -0.9799509    4.131355    1.606988    8.093624
#>      lambda[6,5] lambda[7,5] lambda[8,5] lambda[1,6] lambda[2,6] lambda[3,6]
#> [1,]    1.075836   -2.674489   -2.874608   -2.356018   0.4808712    6.140957
#>      lambda[4,6] lambda[5,6] lambda[6,6] lambda[7,6] lambda[8,6] lambda[1,7]
#> [1,]    2.902257    9.147293    1.076019   -2.675102   -2.874025   -1.590093
#>      lambda[2,7] lambda[3,7] lambda[4,7] lambda[5,7] lambda[6,7] lambda[7,7]
#> [1,]   -1.407007    3.452401    2.646577    9.146409    1.076011   -2.675986
#>      lambda[8,7] lambda[1,8] lambda[2,8] lambda[3,8] lambda[4,8] lambda[5,8]
#> [1,]   -2.874675    1.694752  -0.3035066    7.609011    2.645576    9.147265
#>      lambda[6,8] lambda[7,8] lambda[8,8] lambda[1,9] lambda[2,9] lambda[3,9]
#> [1,]    1.075141   -2.676518   -2.874493    2.255487   -2.003646    7.610083
#>      lambda[4,9] lambda[5,9] lambda[6,9] lambda[7,9] lambda[8,9] lambda[1,10]
#> [1,]    2.646172    9.146973    1.076338   -2.677366    -2.87353     4.617292
#>      lambda[2,10] lambda[3,10] lambda[4,10] lambda[5,10] lambda[6,10]
#> [1,]    -2.002528     7.609622     2.645527     9.147102     1.076642
#>      lambda[7,10] lambda[8,10] lambda_mean[1] lambda_mean[2] lambda_mean[3]
#> [1,]    -2.676585    -2.872543     -0.2756945     -0.6702948     -0.5776078
#>      lambda_mean[4] lambda_mean[5] lambda_mean[6] lambda_mean[7] lambda_mean[8]
#> [1,]     -0.6736658     -0.2367367     -0.1730657     -0.1640918     0.06107859
#>      lambda_mean[9] lambda_mean[10] lambda_mean[11] lambda_mean[12]
#> [1,]      -1.941634       -1.209127          1.0512       -1.285199
#>      lambda_mean[13] lambda_mean[14] lambda_mean[15] lambda_mean[16]
#> [1,]      -0.8763456       0.9417437        1.202447       -2.146079
#>      lambda_mean[17] lambda_mean[18] lambda_mean[19] lambda_mean[20]
#> [1,]       -1.801285        1.847777        1.707896        1.458258
#>      lambda_mean[21] lambda_mean[22] lambda_mean[23] lambda_mean[24]
#> [1,]        3.327993       -1.139231       -1.784976       -2.874715
#>      lambda_mean[25] lambda_mean[26] lambda_mean[27] lambda_mean[28]
#> [1,]       0.4036746        1.979693        4.892957       0.2423914
#>      lambda_mean[29] lambda_mean[30] lambda_mean[31] lambda_mean[32]
#> [1,]        4.638002       0.6130669        -2.67486        -1.90782
#>      lambda_mean[33] lambda_mean[34] lambda_mean[35] lambda_mean[36]
#> [1,]      -0.9799509        4.131355        1.606988        8.093624
#>      lambda_mean[37] lambda_mean[38] lambda_mean[39] lambda_mean[40]
#> [1,]        1.075836       -2.356018       0.4808712        6.140957
#>      lambda_mean[41] lambda_mean[42] lambda_mean[43] lambda_mean[44]
#> [1,]        2.902257        9.147293       -1.590093       -1.407007
#>      lambda_mean[45] lambda_mean[46] lambda_mean[47] lambda_mean[48]
#> [1,]        3.452401        2.646577        1.694752      -0.3035066
#>      lambda_mean[49] lambda_mean[50] lambda_mean[51] lambda_mean[52]    lprior
#> [1,]        7.609011        2.255487       -2.003646        4.617292 -172.6217
#> 
#>