Nowcasting

Function that uses the [rstan::sampling()] engine to generate nowcasts.

Usage

nowcast(
  .disease_data,
  onset_date,
  report_date,
  strata = NULL,
  dist = c("NegativeBinomial", "Poisson"),
  now = NULL,
  units = NULL,
  max_delay = Inf,
  prior_only = FALSE,
  init = 0,
  proportion_reported = 1,
  ...
)

Arguments

.disease_data

A time series of reporting data in aggregated line list format such that each row has a column for onset date, report date, and

onset_date

In quotations, the name of the column of datatype Date designating the date of case onset. e.g. "onset_week"

report_date

In quotations, the name of the column of datatype Date designating the date of case report. e.g. "report_week"

strata

Character vector of names of the strata included in the data.

dist

Distribution. Either "NegativeBinomial" or "Poisson"

now

An object of datatype Date indicating the date at which to perform the nowcast.

units

Time scale of reporting. Options: "1 day", "1 week".

max_delay

Maximum possible delay observed or considered for estimation of the delay distribution (numeric). Default: `Inf`

prior_only

Boolean variable indicating whether to compute only the prior

init

Initial values for [rstan::sampling()]

proportion_reported

A decimal greater than 0 and less than or equal to 1 representing the proportion of all cases expected to be reported. Default: 1, e.g. 100 percent of all cases will eventually be reported. For asymptomatic diseases where not all cases will ever be reported, or for outbreaks in which severe under-reporting is expected, change this to less than 1.

...

Additional arguments to pass to [rstan::sampling()]

Examples

# Load the data
data(denguedat)
now <- as.Date("1990-10-01")

# Run a nowcast with very few iterations
nowcast(denguedat, "onset_week", "report_week", iter = 10, init = 0)
#> ℹ Computing a nowcast for 2010-11-29 per weeks
#> ℹ Assuming data is linelist-data where each observation is a test. If you are working with count-data set `data_type = "count"`
#> 
#> SAMPLING FOR MODEL 'nowcast' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.009736 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 97.36 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: No variance estimation is
#> Chain 1:          performed for num_warmup < 20
#> Chain 1: 
#> Chain 1: Iteration: 1 / 10 [ 10%]  (Warmup)
#> Chain 1: Iteration: 2 / 10 [ 20%]  (Warmup)
#> Chain 1: Iteration: 3 / 10 [ 30%]  (Warmup)
#> Chain 1: Iteration: 4 / 10 [ 40%]  (Warmup)
#> Chain 1: Iteration: 5 / 10 [ 50%]  (Warmup)
#> Chain 1: Iteration: 6 / 10 [ 60%]  (Sampling)
#> Chain 1: Iteration: 7 / 10 [ 70%]  (Sampling)
#> Chain 1: Iteration: 8 / 10 [ 80%]  (Sampling)
#> Chain 1: Iteration: 9 / 10 [ 90%]  (Sampling)
#> Chain 1: Iteration: 10 / 10 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.121 seconds (Warm-up)
#> Chain 1:                0.139 seconds (Sampling)
#> Chain 1:                0.26 seconds (Total)
#> Chain 1: 
#> Error : Exception: neg_binomial_2_log_rng: Random number that came from gamma distribution is 1.97542e+09, but must be less than 1073741824.000000 (in 'nowcast', line 159, column 10 to column 100)
#>  [1] "Chain 1: Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:"                        
#>  [2] "Chain 1: Exception: gamma_lpdf: Random variable[1] is 0, but must be positive finite! (in 'nowcast', line 134, column 4 to column 76)"          
#>  [3] "Chain 1: If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,"
#>  [4] "Chain 1: but if this warning occurs often then your model may be either severely ill-conditioned or misspecified."                              
#>  [5] "Chain 1: "                                                                                                                                      
#>  [6] "Chain 1: Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:"                        
#>  [7] "Chain 1: Exception: gamma_lpdf: Random variable[1] is 0, but must be positive finite! (in 'nowcast', line 134, column 4 to column 76)"          
#>  [8] "Chain 1: If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,"
#>  [9] "Chain 1: but if this warning occurs often then your model may be either severely ill-conditioned or misspecified."                              
#> [10] "Chain 1: "                                                                                                                                      
#> [11] "Chain 1: Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:"                        
#> [12] "Chain 1: Exception: gamma_lpdf: Random variable[1] is 0, but must be positive finite! (in 'nowcast', line 134, column 4 to column 76)"          
#> [13] "Chain 1: If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,"
#> [14] "Chain 1: but if this warning occurs often then your model may be either severely ill-conditioned or misspecified."                              
#> [15] "Chain 1: "                                                                                                                                      
#> [16] "Chain 1: Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:"                        
#> [17] "Chain 1: Exception: gamma_lpdf: Random variable[1] is 0, but must be positive finite! (in 'nowcast', line 134, column 4 to column 76)"          
#> [18] "Chain 1: If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,"
#> [19] "Chain 1: but if this warning occurs often then your model may be either severely ill-conditioned or misspecified."                              
#> [20] "Chain 1: "                                                                                                                                      
#> [21] "Chain 1: Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:"                        
#> [22] "Chain 1: Exception: gamma_lpdf: Random variable[1] is 0, but must be positive finite! (in 'nowcast', line 134, column 4 to column 76)"          
#> [23] "Chain 1: If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,"
#> [24] "Chain 1: but if this warning occurs often then your model may be either severely ill-conditioned or misspecified."                              
#> [25] "Chain 1: "                                                                                                                                      
#> [26] "Chain 1: Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:"                        
#> [27] "Chain 1: Exception: gamma_lpdf: Random variable[1] is 0, but must be positive finite! (in 'nowcast', line 134, column 4 to column 76)"          
#> [28] "Chain 1: If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,"
#> [29] "Chain 1: but if this warning occurs often then your model may be either severely ill-conditioned or misspecified."                              
#> [30] "Chain 1: "                                                                                                                                      
#> [31] "Chain 1: Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:"                        
#> [32] "Chain 1: Exception: gamma_lpdf: Random variable[1] is 0, but must be positive finite! (in 'nowcast', line 134, column 4 to column 76)"          
#> [33] "Chain 1: If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,"
#> [34] "Chain 1: but if this warning occurs often then your model may be either severely ill-conditioned or misspecified."                              
#> [35] "Chain 1: "                                                                                                                                      
#> [36] "Chain 1: Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:"                        
#> [37] "Chain 1: Exception: gamma_lpdf: Random variable[1] is 0, but must be positive finite! (in 'nowcast', line 134, column 4 to column 76)"          
#> [38] "Chain 1: If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,"
#> [39] "Chain 1: but if this warning occurs often then your model may be either severely ill-conditioned or misspecified."                              
#> [40] "Chain 1: "                                                                                                                                      
#> [41] "Chain 1: Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:"                        
#> [42] "Chain 1: Exception: gamma_lpdf: Random variable[1] is 0, but must be positive finite! (in 'nowcast', line 134, column 4 to column 76)"          
#> [43] "Chain 1: If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,"
#> [44] "Chain 1: but if this warning occurs often then your model may be either severely ill-conditioned or misspecified."                              
#> [45] "Chain 1: "                                                                                                                                      
#> error occurred during calling the sampler; sampling not done
#> Stan model 'nowcast' does not contain samples.