Nowcasting with the `rstan` engine

Nowcasting with the `rstan` engine

Usage

nowcast.rstan(
  .disease_data,
  onset_date,
  report_date,
  num_steps,
  num_delays,
  num_strata,
  dist,
  prior_only,
  control,
  refresh,
  method,
  priors,
  ...
)

Arguments

.disease_data

A time series of reporting data in aggregated line list format such that each row has a column for onset date, report date, and

onset_date

In quotations, the name of the column of datatype Date designating the date of case onset. e.g. "onset_week"

report_date

In quotations, the name of the column of datatype Date designating the date of case report. e.g. "report_week"

num_steps

Number of time steps to run in the model

num_delays

Number of delays to consider in the model

num_strata

Number of strata in the model

dist

Distribution. Either "NegativeBinomial" or "Poisson"

prior_only

Boolean variable indicating whether to compute only the prior distribution

control

Control parameter for [rstan::sampling()]

refresh

Refresh parameter for [rstan::sampling()]

method

Fitting method either `sampling` (recommended for inference), `variational` (recommended for testing) or `optimization`. The `sampling` method calls [rstan::sampling()] while the `variational` calls [rstan::vb()] and `optimization` calls [rstan::optimizing()]

priors

A list of all of the nowcast priors. You can use [set_priors()] to change the priors of the function (see details)

...

Additional arguments to pass to [rstan::sampling()]