Get nowcast priors

Function to return the default values for the [nowcast()] function as a list.

Usage

set_priors(
  mu_degree = 1,
  nu_degree = 1,
  mu_is_constant = FALSE,
  nu_is_constant = TRUE,
  mu_sd_prior = "standard_normal",
  nu_sd_prior = "standard_normal",
  mu_sd_param_1 = 0,
  mu_sd_param_2 = 0.1,
  nu_sd_param_1 = 0,
  nu_sd_param_2 = 0.1,
  mu_0_mean_param_1 = "auto",
  mu_0_mean_param_2 = 0.01,
  mu_0_sd_param_1 = "auto",
  mu_0_sd_param_2 = 0.01,
  nu_0_mean_param_1 = 0,
  nu_0_mean_param_2 = 0.01,
  nu_0_sd_param_1 = 0,
  nu_0_sd_param_2 = 0.01,
  mu_0_mean_hyperprior = "standard_normal",
  nu_0_mean_hyperprior = "standard_normal",
  mu_0_sd_hyperprior = "standard_normal",
  nu_0_sd_hyperprior = "standard_normal",
  r_prior = "standard_normal",
  r_param_1 = 0,
  r_param_2 = 1
)

Arguments

mu_degree

Integer. Degree of the epidemic trend.

nu_degree

Integer. Degree of the delay trend.

mu_is_constant

Logical. Indicates whether the epidemic trend is constant.

nu_is_constant

Logical. Indicates whether the delay trend is constant.

mu_sd_prior

Character. Prior for the epidemic trend error.

nu_sd_prior

Character. Prior for the delay trend error.

mu_sd_param_1

Numeric. First parameter for the epidemic trend error.

mu_sd_param_2

Numeric. Second parameter for the epidemic trend error.

nu_sd_param_1

Numeric. First parameter for the delay trend error.

nu_sd_param_2

Numeric. Second parameter for the delay trend error.

mu_0_mean_param_1

Character. Prior for the initial epidemic trend mean's first parameter.

mu_0_mean_param_2

Character. Prior for the initial epidemic trend mean's second parameter.

mu_0_sd_param_1

Character. Prior for the initial epidemic trend standard deviation's first parameter.

mu_0_sd_param_2

Character. Prior for the initial epidemic trend standard deviation's second parameter.

nu_0_mean_param_1

Character. Prior for the initial delay trend mean's first parameter.

nu_0_mean_param_2

Character. Prior for the initial delay trend mean's second parameter.

nu_0_sd_param_1

Character. Prior for the initial delay trend standard deviation's first parameter.

nu_0_sd_param_2

Character. Prior for the initial delay trend standard deviation's second parameter.

mu_0_mean_hyperprior

Prior distribution for the mean of the epidemic trend.

nu_0_mean_hyperprior

Prior distribution for the mean of the delay trend. Default is normal.

mu_0_sd_hyperprior

Prior distribution for the standard deviation of the epidemic trend.

nu_0_sd_hyperprior

Prior distribution for the standard deviation of the delay trend.

r_prior

Character. Prior for the negative binomial precision parameter.

r_param_1

Numeric. First parameter for the dispersion prior if negative binomial.

r_param_2

Numeric. Second parameter for the dispersion prior if negative binomial.

Value

A list with all the priors for the [nowcast()] function.

Examples

#Get the default priors
set_priors()
#> $mu_degree
#> [1] 1
#> 
#> $nu_degree
#> [1] 1
#> 
#> $mu_is_constant
#> [1] FALSE
#> 
#> $nu_is_constant
#> [1] TRUE
#> 
#> $mu_sd_prior
#> [1] "standard_normal"
#> 
#> $nu_sd_prior
#> [1] "standard_normal"
#> 
#> $mu_sd_param_1
#> [1] 0
#> 
#> $mu_sd_param_2
#> [1] 0.1
#> 
#> $nu_sd_param_1
#> [1] 0
#> 
#> $nu_sd_param_2
#> [1] 0.1
#> 
#> $mu_0_mean_param_1
#> [1] "auto"
#> 
#> $mu_0_mean_param_2
#> [1] 0.01
#> 
#> $mu_0_sd_param_1
#> [1] "auto"
#> 
#> $mu_0_sd_param_2
#> [1] 0.01
#> 
#> $nu_0_mean_param_1
#> [1] 0
#> 
#> $nu_0_mean_param_2
#> [1] 0.01
#> 
#> $nu_0_sd_param_1
#> [1] 0
#> 
#> $nu_0_sd_param_2
#> [1] 0.01
#> 
#> $mu_0_mean_hyperprior
#> [1] "standard_normal"
#> 
#> $nu_0_mean_hyperprior
#> [1] "standard_normal"
#> 
#> $mu_0_sd_hyperprior
#> [1] "standard_normal"
#> 
#> $nu_0_sd_hyperprior
#> [1] "standard_normal"
#> 
#> $r_prior
#> [1] "standard_normal"
#> 
#> $r_param_1
#> [1] 0
#> 
#> $r_param_2
#> [1] 1
#> 

#Change the distribution of the priors sd for the epidemic trend as well as the degree of the delays
set_priors(nu_degree = 2, mu_sd_prior = "cauchy")
#> $mu_degree
#> [1] 1
#> 
#> $nu_degree
#> [1] 2
#> 
#> $mu_is_constant
#> [1] FALSE
#> 
#> $nu_is_constant
#> [1] TRUE
#> 
#> $mu_sd_prior
#> [1] "cauchy"
#> 
#> $nu_sd_prior
#> [1] "standard_normal"
#> 
#> $mu_sd_param_1
#> [1] 0
#> 
#> $mu_sd_param_2
#> [1] 0.1
#> 
#> $nu_sd_param_1
#> [1] 0
#> 
#> $nu_sd_param_2
#> [1] 0.1
#> 
#> $mu_0_mean_param_1
#> [1] "auto"
#> 
#> $mu_0_mean_param_2
#> [1] 0.01
#> 
#> $mu_0_sd_param_1
#> [1] "auto"
#> 
#> $mu_0_sd_param_2
#> [1] 0.01
#> 
#> $nu_0_mean_param_1
#> [1] 0
#> 
#> $nu_0_mean_param_2
#> [1] 0.01
#> 
#> $nu_0_sd_param_1
#> [1] 0
#> 
#> $nu_0_sd_param_2
#> [1] 0.01
#> 
#> $mu_0_mean_hyperprior
#> [1] "standard_normal"
#> 
#> $nu_0_mean_hyperprior
#> [1] "standard_normal"
#> 
#> $mu_0_sd_hyperprior
#> [1] "standard_normal"
#> 
#> $nu_0_sd_hyperprior
#> [1] "standard_normal"
#> 
#> $r_prior
#> [1] "standard_normal"
#> 
#> $r_param_1
#> [1] 0
#> 
#> $r_param_2
#> [1] 1
#>