Simulate a process

This function simulates a state-space process for testing purposes in a model with multiple delays and strata. It estimates the process over a specified number of steps, with options for trend degrees, priors, and error distributions.

Usage

simulate_process_for_testing(
  num_steps = 10,
  num_delays = 8,
  num_strata = 2,
  mu_degree = 2,
  nu_degree = 1,
  mu_is_constant = FALSE,
  nu_is_constant = TRUE,
  mu_error_prior = "normal",
  nu_error_prior = "normal",
  mu_param_1 = 0,
  mu_param_2 = 0.1,
  nu_param_1 = 0,
  nu_param_2 = 0.1,
  mu_0_prior = "normal",
  nu_0_prior = "normal",
  mu_0_param_1 = log(100),
  mu_0_param_2 = 0.01,
  nu_0_param_1 = log(6),
  nu_0_param_2 = 0.01,
  r_prior = "normal",
  is_negative_binomial = TRUE,
  r_param_1 = 0,
  r_param_2 = 1
)

Arguments

num_steps

Integer. Number of time steps to simulate. Default is 10.

num_delays

Integer. Number of delay strata. Default is 8.

num_strata

Integer. Number of strata for the population. Default is 2.

mu_degree

Integer. Degree of the epidemic trend. Default is 2.

nu_degree

Integer. Degree of the delay trend. Default is 1.

mu_is_constant

Logical. Indicates whether the epidemic trend is constant. Default is FALSE.

nu_is_constant

Logical. Indicates whether the delay trend is constant. Default is TRUE.

mu_error_prior

Character. Prior for the epidemic trend error. Default is "normal".

nu_error_prior

Character. Prior for the delay trend error. Default is "normal".

mu_param_1

Numeric. First parameter for the epidemic trend error. Default is 0.0.

mu_param_2

Numeric. Second parameter for the epidemic trend error. Default is 1.0.

nu_param_1

Numeric. First parameter for the delay trend error. Default is 0.0.

nu_param_2

Numeric. Second parameter for the delay trend error. Default is 1.0.

mu_0_prior

Character. Prior for the initial epidemic trend distribution. Default is "normal".

nu_0_prior

Character. Prior for the initial delay trend distribution. Default is "normal".

mu_0_param_1

Numeric. First parameter for the initial epidemic trend distribution. Default is 0.0.

mu_0_param_2

Numeric. Second parameter for the initial epidemic trend distribution. Default is 1.0.

nu_0_param_1

Numeric. First parameter for the initial delay trend distribution. Default is 0.0.

nu_0_param_2

Numeric. Second parameter for the initial delay trend distribution. Default is 1.0.

r_prior

Character. Prior for the negative binomial precision parameter. Default is "normal".

is_negative_binomial

Logical. Indicates whether the data is modeled using a negative binomial distribution. Default is TRUE.

r_param_1

Numeric. First parameter for the dispersion prior if negative binomial. Default is 0.0.

r_param_2

Numeric. Second parameter for the dispersion prior if negative binomial. Default is 1.0.

Value

A tibble with simulated state-space process results, including the onset and report dates, strata, delays, and the number of observed cases per time step.

Examples

simulate_process_for_testing(num_steps = 20, num_delays = 5, num_strata = 3)
#> # A tibble: 300 × 7
#>    .strata .delay .tval     n lambda onset_date report_date
#>      <int>  <int> <dbl> <dbl>  <dbl> <date>     <date>     
#>  1       1      1     1   928   605. 2024-08-23 2024-08-23 
#>  2       1      1     2   353   661. 2024-08-24 2024-08-24 
#>  3       1      1     3   216   656. 2024-08-25 2024-08-25 
#>  4       1      1     4  1733   618. 2024-08-26 2024-08-26 
#>  5       1      1     5   187   530. 2024-08-27 2024-08-27 
#>  6       1      1     6  1714   392. 2024-08-28 2024-08-28 
#>  7       1      1     7    31   285. 2024-08-29 2024-08-29 
#>  8       1      1     8   231   243. 2024-08-30 2024-08-30 
#>  9       1      1     9   344   229. 2024-08-31 2024-08-31 
#> 10       1      1    10    31   193. 2024-09-01 2024-09-01 
#> # ℹ 290 more rows